% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_database.R
\name{add_database}
\alias{add_database}
\title{Add PEcAn database information to settings list}
\usage{
add_database(settings, host = "postgres", user = "bety",
  password = "bety", dbname = "bety", driver = "PostgreSQL",
  dbfiles = "/data/dbfiles/", write = TRUE, overwrite = FALSE, ...)
}
\arguments{
\item{settings}{Input settings list (list)}

\item{host}{Database server hostname (character, default =
"postgres", so it works with Docker)}

\item{user}{Database user name (character, default = "bety")}

\item{password}{Database password (character, default = "bety")}

\item{dbname}{Database name (character, default = "bety")}

\item{driver}{Database driver (character, default = "PostgreSQL")}

\item{dbfiles}{Path to \code{dbfiles} directory (character, default = "/data/dbfiles")}

\item{write}{Whether or not to write to the database (logical,
default = \code{TRUE})}

\item{overwrite}{Whether or not to overwrite any already existing
input settings (logical, default = \code{FALSE})}

\item{...}{Additional named PEcAn database configuration tags}
}
\value{
Updated settings list with database information included
}
\description{
Add PEcAn database information to settings list
}
\author{
Alexey Shiklomanov
}
