% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pft.R
\name{add_pft}
\alias{add_pft}
\alias{add_pft_list}
\title{Add a PFT or list of PFTs to a settings object}
\usage{
add_pft(settings, name, ...)

add_pft_list(settings, pfts, ...)
}
\arguments{
\item{settings}{Input PEcAn settings list}

\item{name}{PFT name (character)}

\item{...}{Additional arguments for modifying the PFT.}

\item{pfts}{Either a character vector of PFT names, or a list of
PFT objects (which must each include an element named "name")}
}
\value{
Updated settings list with PFTs added
}
\description{
Add a PFT or list of PFTs to a settings object
}
\examples{
settings <- list()
add_pft(settings, "Optics.Temperate_Early_Hardwood")
add_pft_list(settings, sprintf("Temperate_\%s_Hardwood", c("Early", "Mid", "Late")))
add_pft_list(
  settings,
  list(list(name = "deciduous", my_number = 3),
       list(name = "coniferous", my_number = 6))
)  
if (require("magrittr")) {
  list() \%>\%
    add_pft("early_hardwood") \%>\%
    add_pft("mid_hardwood") \%>\%
    add_pft("late_hardwood")
}
}
\author{
Alexey Shiklomanov
}
