% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_api.R
\name{add_plumber_definition}
\alias{add_plumber_definition}
\title{Mount plumber definition}
\usage{
add_plumber_definition(pr, endpoint, file, ...)
}
\arguments{
\item{pr}{The host Plumber router.}

\item{endpoint}{route endpoint}

\item{file}{Path to file to plumb}

\item{...}{arguments passed to plumber::pr}
}
\value{
a plumber object with additional mounted routes
}
\description{
Mount a new set of routes to a plumber object from a file.
}
