% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rabbitmq.R
\name{add_rabbitmq}
\alias{add_rabbitmq}
\title{Add RabbitMQ configuration}
\usage{
add_rabbitmq(settings, model_queue = NULL, con = NULL,
  rabbitmq_user = "guest", rabbitmq_password = "guest",
  rabbitmq_service = "rabbitmq", rabbitmq_service_port = 5672,
  rabbitmq_vhost = "\%2f", overwrite = FALSE)
}
\arguments{
\item{settings}{Partially completed PEcAn \code{settings} list.}

\item{model_queue}{Name of RabbitMQ model queue (character, default
= \code{NULL}). This should be in the form \code{modelname_modelrevision}.
If this is \code{NULL}, this function will try to figure it out based
on the model ID in the settings object, which requires access to
the database (i.e. \code{con} must not be \code{NULL}).}

\item{con}{Database connection object (default = \code{NULL}). Ignored
unless \code{model_queue} is \code{NULL}.}

\item{rabbitmq_user}{Username for RabbitMQ server (character,
default = "guest")}

\item{rabbitmq_password}{Password for RabbitMQ server (character,
default = "guest")}

\item{rabbitmq_service}{Name of RabbitMQ \code{docker-compose} service
(character, default = "rabbitmq")}

\item{rabbitmq_service_port}{RabbitMQ service port (numeric or
character, default = 5672). Note that this is internal to the
Docker stack, \emph{not}. The only reason this should be changed is if you
changed low-level RabbitMQ sett}

\item{rabbitmq_vhost}{RabbitMQ vhost (character, default = "%2f").
The only reason this should be changed is if you change the
low-level RabbitMQ setup in the \code{docker-compose.yml} file.}

\item{overwrite}{Whether or not to overwrite existing \code{settings}
tags (logical; default = \code{FALSE})}
}
\value{
Modified settings list with RabbitMQ configuration added.
}
\description{
Add RabbitMQ configuration
}
\author{
Alexey Shiklomanov
}
