% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_workflow.R
\name{add_workflow}
\alias{add_workflow}
\title{Add information from workflow data frame into settings list.}
\usage{
add_workflow(settings, workflow_df, overwrite = FALSE)
}
\arguments{
\item{settings}{Partially completed PEcAn \code{settings} list.}

\item{workflow_df}{Workflow \code{data.frame}, such as that returned by
\link{insert_new_workflow}, or from running query \verb{SELECT * FROM workflows}}

\item{overwrite}{Whether or not to overwrite existing \code{settings}
tags (logical; default = \code{FALSE})}
}
\value{
PEcAn settings list with workflow information added
}
\description{
Add information from workflow data frame into settings list.
}
\author{
Alexey Shiklomanov
}
