% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Perform a sentiment analysis of a parsed, tidy talk page}
\usage{
analyze(parsed_talk, lexicon, .silent = TRUE)
}
\arguments{
\item{lexicon}{lexicon object provided by \link[tidytext:get_sentiments]{tidytext::get_sentiments}}

\item{.silent}{suppresses messages}

\item{tidy_page}{the output from \link{parse_discussion}}
}
\description{
Uses an emotion lexicon to output a breakdown of sentiments
expressed.
}
