% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{Consistent anonymization}
\usage{
anonymize(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Uses HMAC (hashed message authentication code) hashing with the
name of the machine as the key, enabling consistent username anonymization
within the lifetime of the machine running the code. This means the output
from several talk pages that have authors in common will have hashes in
common as well.
}
\seealso{
\link[openssl:hashing]{openssl::hashing}
}
