\name{API-package}
\alias{API-package}
\alias{API}
\docType{package}
\title{
\packageTitle{API}
}
\description{
\packageDescription{API}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{API}
\packageIndices{API}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{API}

Maintainer: \packageMaintainer{API}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
