% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{children}
\alias{children}
\title{Gets a new grouped symbol whose output contains inputs to output nodes of the original symbol.}
\usage{
children(x)
}
\arguments{
\item{x}{The input symbol}
}
\description{
Gets a new grouped symbol whose output contains inputs to output nodes of the original symbol.
}

