% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_report.R
\name{pdf_report_gen}
\alias{pdf_report_gen}
\alias{csv_report}
\title{Generate reports}
\usage{
pdf_report_gen(req)

csv_report(req)
}
\arguments{
\item{req}{rook request object}
}
\description{
API endpoints
}
