% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds_fileserver.R
\name{dbfile_url}
\alias{dbfile_url}
\title{Build a THREDDS fileServer URL for a specific dbfile}
\usage{
dbfile_url(target, ...)
}
\arguments{
\item{target}{Target file path, relative to \code{dbfiles} root folder (character)}

\item{...}{Additional arguments to \link{thredds_fs_url}}
}
\value{
THREDDS HTTP fileServer URL to dbfile (character)
}
\description{
Build a THREDDS fileServer URL for a specific dbfile
}
\author{
Alexey Shiklomanov
}
