% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_id.R
\name{get_model_id}
\alias{get_model_id}
\title{Retrieve database ID of a particular version of a model}
\usage{
get_model_id(con, name, revision, multi_action = "last")
}
\arguments{
\item{con}{Database connection object (Pqconnection)}

\item{name}{Model name (character)}

\item{revision}{Model version/revision (character)}

\item{multi_action}{Action to take if multiple models found
(character). Must be one of "first", "last" (default), "all", or "error".}
}
\value{
Model ID, as \code{integer64}
}
\description{
Retrieve database ID of a particular version of a model
}
\author{
Alexey Shiklomanov
}
