% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_workflow.R
\name{get_next_workflow_id}
\alias{get_next_workflow_id}
\title{Get next workflow ID from sequence. Note that this also updates the
workflow sequence.}
\usage{
get_next_workflow_id(con)
}
\arguments{
\item{con}{Database connection, as created by \link[RPostgres:dbConnect]{RPostgres::dbConnect}}
}
\value{
Workflow ID, as numeric/base64 integer
}
\description{
Get next workflow ID from sequence. Note that this also updates the
workflow sequence.
}
\author{
Alexey Shiklomanov
}
