% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds_fileserver.R
\name{get_run_id}
\alias{get_run_id}
\title{Get a run ID from the \code{runs.txt} file}
\usage{
get_run_id(workflow_id, ...)
}
\arguments{
\item{workflow_id}{ID of target workflow (numeric or character)}

\item{...}{Additional arguments to \link{thredds_fs_url}}
}
\value{
A single output ID (integer64)
}
\description{
Get a run ID from the \code{runs.txt} file
}
\author{
Alexey Shiklomanov
}
