% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.graph.R
\name{graph.viz}
\alias{graph.viz}
\title{Convert symbol to Graphviz or visNetwork visualisation.}
\usage{
graph.viz(symbol, shape = NULL, direction = "TD", type = "graph",
  graph.width.px = NULL, graph.height.px = NULL)
}
\arguments{
\item{symbol}{a \code{string} representing the symbol of a model.}

\item{shape}{a \code{numeric} representing the input dimensions to the symbol.}

\item{direction}{a \code{string} representing the direction of the graph, either TD or LR.}

\item{type}{a \code{string} representing the rendering engine of the the graph, either graph or vis.}

\item{graph.width.px}{a \code{numeric} representing the size (width) of the graph. In pixels}

\item{graph.height.px}{a \code{numeric} representing the size (height) of the graph. In pixels}
}
\value{
a graph object ready to be displayed with the \code{print} function.
}
\description{
Convert symbol to Graphviz or visNetwork visualisation.
}

