% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{im2rec}
\alias{im2rec}
\title{Convert images into image recordio format}
\usage{
im2rec(image_lst, root, output_rec, label_width = 1L, pack_label = 0L,
  new_size = -1L, nsplit = 1L, partid = 0L, center_crop = 0L,
  quality = 95L, color_mode = 1L, unchanged = 0L, inter_method = 1L,
  encoding = ".jpg")
}
\arguments{
\item{image_lst}{The image lst file}

\item{root}{The root folder for image files}

\item{output_rec}{The output rec file}

\item{label_width}{The label width in the list file. Default is 1.}

\item{pack_label}{Whether to also pack multi dimenional label in the record file. Default is 0.}

\item{new_size}{The shorter edge of image will be resized to the newsize. 
Original images will be packed by default.}

\item{nsplit}{It is used for part generation, logically split the image.lst to NSPLIT parts by position.
Default is 1.}

\item{partid}{It is used for part generation, pack the images from the specific part in image.lst.
Default is 0.}

\item{center_crop}{Whether to crop the center image to make it square. Default is 0.}

\item{quality}{JPEG quality for encoding (1-100, default: 95) or PNG compression for encoding (1-9, default: 3).}

\item{color_mode}{Force color (1), gray image (0) or keep source unchanged (-1). Default is 1.}

\item{unchanged}{Keep the original image encoding, size and color. If set to 1, it will ignore the others parameters.}

\item{inter_method}{NN(0), BILINEAR(1), CUBIC(2), AREA(3), LANCZOS4(4), AUTO(9), RAND(10). Default is 1.}

\item{encoding}{The encoding type for images. It can be '.jpg' or '.png'. Default is '.jpg'.}
}
\description{
Convert images into image recordio format
}

