% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{internals}
\alias{internals}
\title{Get a symbol that contains all the internals}
\usage{
internals(x)
}
\arguments{
\item{x}{The input symbol}
}
\description{
Get a symbol that contains all the internals
}

