% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{is.serialized}
\alias{is.serialized}
\title{Check if the model has been serialized into RData-compatiable format.}
\usage{
is.serialized(model)
}
\value{
Logical indicator
}
\description{
Check if the model has been serialized into RData-compatiable format.
}

