% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_runs.R, R/output_url.R
\name{list_runs}
\alias{list_runs}
\title{List all runs associated with a particular workflow}
\usage{
list_runs(con, workflow_id)

list_runs(con, workflow_id)
}
\arguments{
\item{con}{Database connection, as created by \link[RPostgres:dbConnect]{RPostgres::dbConnect}}

\item{workflow_id}{ID of target workflow (character or numeric)}

\item{workflow_id}{ID of target workflow (character or numeric)}
}
\value{
Runs \code{data.frame} subset to rows containing specific workflow

Runs \code{data.frame} subset to rows containing specific workflow
}
\description{
List all runs associated with a particular workflow

List all runs associated with a particular workflow
}
\author{
Alexey Shiklomanov

Alexey Shiklomanov
}
