% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_workflow.R
\name{listToXML}
\alias{listToXML}
\title{Convert List to XML}
\usage{
listToXML(item, tag)
}
\arguments{
\item{item}{object to be converted. Despite the function name, need not actually be a list}

\item{tag}{xml tag}
}
\value{
xmlNode
}
\description{
Can convert list or other object to an xml object using xmlNode
}
\author{
David LeBauer, Carl Davidson, Rob Kooper
}
