% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{mx.ctx.default}
\alias{mx.ctx.default}
\title{Set/Get default context for array creation.}
\usage{
mx.ctx.default(new = NULL)
}
\arguments{
\item{new, }{optional takes \code{mx.cpu()} or \code{mx.gpu(id)}, new default ctx.}
}
\value{
The default context.
}
\description{
Set/Get default context for array creation.
}

