% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executor.R
\name{mx.exec.backward}
\alias{mx.exec.backward}
\title{Peform an backward on the executors
This function will MUTATE the state of exec}
\usage{
mx.exec.backward(exec, ...)
}
\description{
Peform an backward on the executors
This function will MUTATE the state of exec
}

