% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn.infer.R
\name{mx.infer.rnn.one}
\alias{mx.infer.rnn.one}
\title{Inference for one-to-one fusedRNN (CUDA) models}
\usage{
mx.infer.rnn.one(infer.data, symbol, arg.params, aux.params,
  input.params = NULL, ctx = mx.cpu())
}
\arguments{
\item{infer.data}{Data iterator created by mx.io.bucket.iter}

\item{symbol}{Symbol used for inference}

\item{ctx}{}
}
\description{
Inference for one-to-one fusedRNN (CUDA) models
}

