% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn.infer.R
\name{mx.infer.rnn.one.unroll}
\alias{mx.infer.rnn.one.unroll}
\title{Inference for one-to-one unroll models}
\usage{
mx.infer.rnn.one.unroll(infer.data, symbol, num_hidden, arg.params, aux.params,
  init_states = NULL, ctx = mx.cpu())
}
\arguments{
\item{infer.data}{NDArray}

\item{symbol}{Model used for inference}

\item{num_hidden}{}

\item{ctx}{}
}
\description{
Inference for one-to-one unroll models
}

