% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn.infer.R
\name{mx.infer.rnn}
\alias{mx.infer.rnn}
\title{Inference of RNN model}
\usage{
mx.infer.rnn(infer.data, model, ctx = mx.cpu())
}
\arguments{
\item{infer.data}{DataIter}

\item{model}{Model used for inference}

\item{ctx}{}
}
\description{
Inference of RNN model
}

