% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer.R
\name{mx.init.create}
\alias{mx.init.create}
\title{Create initialization of argument  like arg.array}
\usage{
mx.init.create(initializer, shape.array, ctx = NULL, skip.unknown = TRUE)
}
\arguments{
\item{initializer}{The initializer.}

\item{shape.array}{A named list that represents the shape of the weights}

\item{ctx}{mx.context The context of the weights}

\item{skip.unknown}{Whether skip the unknown weight types}
}
\description{
Create initialization of argument  like arg.array
}

