% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mx.io.arrayiter}
\alias{mx.io.arrayiter}
\title{Create MXDataIter compatible iterator from R's array}
\usage{
mx.io.arrayiter(data, label, batch.size = 128, shuffle = FALSE)
}
\arguments{
\item{data}{The data array.}

\item{label}{The label array.}

\item{batch.size}{The batch size used to pack the array.}

\item{shuffle}{Whether shuffle the data}
}
\description{
Create MXDataIter compatible iterator from R's array
}

