% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx.io.bucket.iter.R
\name{mx.io.bucket.iter}
\alias{mx.io.bucket.iter}
\title{Create Bucket Iter}
\usage{
mx.io.bucket.iter(buckets, batch.size, data.mask.element = 0,
  shuffle = FALSE, seed = 123)
}
\arguments{
\item{buckets}{The data array.}

\item{batch.size}{The batch size used to pack the array.}

\item{data.mask.element}{The element to mask}

\item{shuffle}{Whether shuffle the data}

\item{seed}{The random seed}
}
\description{
Create Bucket Iter
}

