% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.io.MNISTIter}
\alias{mx.io.MNISTIter}
\title{Iterating on the MNIST dataset.}
\usage{
mx.io.MNISTIter(...)
}
\arguments{
\item{image}{string, optional, default='./train-images-idx3-ubyte'
Dataset Param: Mnist image path.}

\item{label}{string, optional, default='./train-labels-idx1-ubyte'
Dataset Param: Mnist label path.}

\item{batch.size}{int, optional, default='128'
Batch Param: Batch Size.}

\item{shuffle}{boolean, optional, default=1
Augmentation Param: Whether to shuffle data.}

\item{flat}{boolean, optional, default=0
Augmentation Param: Whether to flat the data into 1D.}

\item{seed}{int, optional, default='0'
Augmentation Param: Random Seed.}

\item{silent}{boolean, optional, default=0
Auxiliary Param: Whether to print out data info.}

\item{num.parts}{int, optional, default='1'
partition the data into multiple parts}

\item{part.index}{int, optional, default='0'
the index of the part will read}

\item{prefetch.buffer}{, optional, default=4
Maximum number of batches to prefetch.}

\item{dtype}{{None, 'float16', 'float32', 'float64', 'int32', 'int64', 'uint8'},optional, default='None'
Output data type. ``None`` means no change.}
}
\value{
iter The result mx.dataiter
}
\description{
One can download the dataset from http://yann.lecun.com/exdb/mnist/
}
\details{
Defined in src/io/iter_mnist.cc:L265
}

