% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.io.SequentialSampler}
\alias{mx.io.SequentialSampler}
\title{Returns the sequential sampler iterator.}
\usage{
mx.io.SequentialSampler(...)
}
\arguments{
\item{length}{long (non-negative), required
Length of the sequence.}

\item{start}{int, optional, default='0'
Start of the index.}

\item{batch.size}{int (non-negative), required
Batch size.}

\item{last.batch}{{'discard', 'keep', 'rollover'},optional, default='keep'
Specifies how the last batch is handled if batch_size does not evenly divide sequence length. If 'keep', the last batch will be returned directly, but will contain less element than `batch_size` requires. If 'discard', the last batch will be discarded. If 'rollover', the remaining elements will be rolled over to the next iteration. Note: legacy batch param with round_batch will always round data in order to always provide full batchs. Rollover behavior will instead result in different iteration sizes for each epoch.}
}
\value{
iter The result mx.dataiter
}
\description{
Defined in src/io/iter_sampler.cc:L97
}
