% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_scheduler.R
\name{mx.lr_scheduler.FactorScheduler}
\alias{mx.lr_scheduler.FactorScheduler}
\title{Learning rate scheduler. Reduction based on a factor value.}
\usage{
mx.lr_scheduler.FactorScheduler(step, factor_val, stop_factor_lr = 1e-08,
  verbose = TRUE)
}
\arguments{
\item{step}{(integer)
Schedule learning rate after n updates}

\item{factor}{(double)
The factor for reducing the learning rate}
}
\value{
scheduler function
}
\description{
Learning rate scheduler. Reduction based on a factor value.
}

