% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.logistic_acc}
\alias{mx.metric.logistic_acc}
\title{Accuracy metric for logistic regression}
\format{An object of class \code{mx.metric} of length 3.}
\usage{
mx.metric.logistic_acc
}
\description{
Accuracy metric for logistic regression
}
\keyword{datasets}

