% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.logloss}
\alias{mx.metric.logloss}
\title{LogLoss metric for logistic regression}
\format{An object of class \code{mx.metric} of length 3.}
\usage{
mx.metric.logloss
}
\description{
LogLoss metric for logistic regression
}
\keyword{datasets}

