% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.Perplexity}
\alias{mx.metric.Perplexity}
\title{Perplexity metric for language model}
\format{An object of class \code{mx.metric} of length 3.}
\usage{
mx.metric.Perplexity
}
\description{
Perplexity metric for language model
}
\keyword{datasets}

