% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{mx.metric.top_k_accuracy}
\alias{mx.metric.top_k_accuracy}
\title{Top-k accuracy metric for classification}
\format{An object of class \code{mx.metric} of length 3.}
\usage{
mx.metric.top_k_accuracy
}
\description{
Top-k accuracy metric for classification
}
\keyword{datasets}

