% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mx.model.init.params}
\alias{mx.model.init.params}
\title{Parameter initialization}
\usage{
mx.model.init.params(symbol, input.shape, output.shape, initializer, ctx)
}
\arguments{
\item{symbol}{The symbolic configuration of the neural network.}

\item{input.shape}{The shape of the input for the neural network.}

\item{output.shape}{The shape of the output for the neural network. It can be NULL.}

\item{initializer, }{initializer object. The initialization scheme for parameters.}

\item{ctx}{mx.context. The devices used to perform initialization.}
}
\description{
Parameter initialization
}

