% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mx.model.save}
\alias{mx.model.save}
\title{Save model checkpoint into file.}
\usage{
mx.model.save(model, prefix, iteration)
}
\arguments{
\item{model}{The feedforward model to be saved.}

\item{prefix}{string prefix of the model name}

\item{iteration}{integer Iteration number of model we would like to load.}
}
\description{
Save model checkpoint into file.
}

