% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.all.finite}
\alias{mx.nd.all.finite}
\title{Check if all the float numbers in the array are finite (used for AMP)}
\arguments{
\item{data}{NDArray
Array}

\item{init.output}{boolean, optional, default=1
Initialize output to 1.}
}
\value{
out The result mx.ndarray
}
\description{
Defined in src/operator/contrib/all_finite.cc:L101
}
