% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.amp.multicast}
\alias{mx.nd.amp.multicast}
\title{Cast function used by AMP, that casts its inputs to the common widest type.}
\arguments{
\item{data}{NDArray-or-Symbol[]
Weights}

\item{num.outputs}{int, required
Number of input/output pairs to be casted to the widest type.}

\item{cast.narrow}{boolean, optional, default=0
Whether to cast to the narrowest type}
}
\value{
out The result mx.ndarray
}
\description{
It casts only between low precision float/FP32 and does not do anything for other types.
}
\details{
Defined in src/operator/tensor/amp_cast.cc:L71
}
