% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.arctanh}
\alias{mx.nd.arctanh}
\title{Returns the element-wise inverse hyperbolic tangent of the input array, \
computed element-wise.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
The storage type of ``arctanh`` output depends upon the input storage type:
}
\details{
- arctanh(default) = default
   - arctanh(row_sparse) = row_sparse
   - arctanh(csr) = csr



Defined in src/operator/tensor/elemwise_unary_op_trig.cc:L281
}

