% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.broadcast.axes}
\alias{mx.nd.broadcast.axes}
\title{Broadcasts the input array over particular axes.}
\arguments{
\item{data}{NDArray-or-Symbol
The input}

\item{axis}{Shape(tuple), optional, default=[]
The axes to perform the broadcasting.}

\item{size}{Shape(tuple), optional, default=[]
Target sizes of the broadcasting axes.}
}
\value{
out The result mx.ndarray
}
\description{
Broadcasting is allowed on axes with size 1, such as from `(2,1,3,1)` to
`(2,8,3,9)`. Elements will be duplicated on the broadcasted axes.
}
\details{
Example::

// given x of shape (1,2,1)
   x = [[[ 1.],
         [ 2.]]]

// broadcast x on on axis 2
   broadcast_axis(x, axis=2, size=3) = [[[ 1.,  1.,  1.],
                                         [ 2.,  2.,  2.]]]
   // broadcast x on on axes 0 and 2
   broadcast_axis(x, axis=(0,2), size=(2,3)) = [[[ 1.,  1.,  1.],
                                                 [ 2.,  2.,  2.]],
                                                [[ 1.,  1.,  1.],
                                                 [ 2.,  2.,  2.]]]

Defined in src/operator/tensor/broadcast_reduce_op_value.cc:L238
}

