% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.broadcast.logical.or}
\alias{mx.nd.broadcast.logical.or}
\title{Returns the result of element-wise **logical or** with broadcasting.}
\arguments{
\item{lhs}{NDArray-or-Symbol
First input to the function}

\item{rhs}{NDArray-or-Symbol
Second input to the function}
}
\value{
out The result mx.ndarray
}
\description{
Example::
}
\details{
x = [[ 1.,  1.,  0.],
        [ 1.,  1.,  0.]]

y = [[ 1.],
        [ 0.]]

broadcast_logical_or(x, y) = [[ 1.,  1.,  1.],
                                 [ 1.,  1.,  0.]]



Defined in src/operator/tensor/elemwise_binary_broadcast_op_logic.cc:L172
}

