% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.broadcast.minimum}
\alias{mx.nd.broadcast.minimum}
\title{Returns element-wise minimum of the input arrays with broadcasting.}
\arguments{
\item{lhs}{NDArray-or-Symbol
First input to the function}

\item{rhs}{NDArray-or-Symbol
Second input to the function}
}
\value{
out The result mx.ndarray
}
\description{
This function compares two input arrays and returns a new array having the element-wise minima.
}
\details{
Example::

x = [[ 1.,  1.,  1.],
        [ 1.,  1.,  1.]]

y = [[ 0.],
        [ 1.]]

broadcast_maximum(x, y) = [[ 0.,  0.,  0.],
                              [ 1.,  1.,  1.]]



Defined in src/operator/tensor/elemwise_binary_broadcast_op_extended.cc:L115
}

