% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.broadcast.plus}
\alias{mx.nd.broadcast.plus}
\title{Returns element-wise sum of the input arrays with broadcasting.}
\arguments{
\item{lhs}{NDArray-or-Symbol
First input to the function}

\item{rhs}{NDArray-or-Symbol
Second input to the function}
}
\value{
out The result mx.ndarray
}
\description{
`broadcast_plus` is an alias to the function `broadcast_add`.
}
\details{
Example::

x = [[ 1.,  1.,  1.],
        [ 1.,  1.,  1.]]

y = [[ 0.],
        [ 1.]]

broadcast_add(x, y) = [[ 1.,  1.,  1.],
                          [ 2.,  2.,  2.]]

broadcast_plus(x, y) = [[ 1.,  1.,  1.],
                           [ 2.,  2.,  2.]]

Supported sparse operations:

broadcast_add(csr, dense(1D)) = dense
   broadcast_add(dense(1D), csr) = dense



Defined in src/operator/tensor/elemwise_binary_broadcast_op_basic.cc:L58
}

