% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.Convolution.v1}
\alias{mx.nd.Convolution.v1}
\title{This operator is DEPRECATED. Apply convolution to input then add a bias.}
\arguments{
\item{data}{NDArray-or-Symbol
Input data to the ConvolutionV1Op.}

\item{weight}{NDArray-or-Symbol
Weight matrix.}

\item{bias}{NDArray-or-Symbol
Bias parameter.}

\item{kernel}{Shape(tuple), required
convolution kernel size: (h, w) or (d, h, w)}

\item{stride}{Shape(tuple), optional, default=[]
convolution stride: (h, w) or (d, h, w)}

\item{dilate}{Shape(tuple), optional, default=[]
convolution dilate: (h, w) or (d, h, w)}

\item{pad}{Shape(tuple), optional, default=[]
pad for convolution: (h, w) or (d, h, w)}

\item{num.filter}{int (non-negative), required
convolution filter(channel) number}

\item{num.group}{int (non-negative), optional, default=1
Number of group partitions. Equivalent to slicing input into num_group
partitions, apply convolution on each, then concatenate the results}

\item{workspace}{long (non-negative), optional, default=1024
Maximum temporary workspace allowed for convolution (MB).This parameter determines the effective batch size of the convolution kernel, which may be smaller than the given batch size. Also, the workspace will be automatically enlarged to make sure that we can run the kernel with batch_size=1}

\item{no.bias}{boolean, optional, default=0
Whether to disable bias parameter.}

\item{cudnn.tune}{{None, 'fastest', 'limited_workspace', 'off'},optional, default='None'
Whether to pick convolution algo by running performance test.
Leads to higher startup time but may give faster speed. Options are:
'off': no tuning
'limited_workspace': run test and pick the fastest algorithm that doesn't exceed workspace limit.
'fastest': pick the fastest algorithm and ignore workspace limit.
If set to None (default), behavior is determined by environment
variable MXNET_CUDNN_AUTOTUNE_DEFAULT: 0 for off,
1 for limited workspace (default), 2 for fastest.}

\item{cudnn.off}{boolean, optional, default=0
Turn off cudnn for this layer.}

\item{layout}{{None, 'NCDHW', 'NCHW', 'NDHWC', 'NHWC'},optional, default='None'
Set layout for input, output and weight. Empty for
default layout: NCHW for 2d and NCDHW for 3d.}
}
\value{
out The result mx.ndarray
}
\description{
This operator is DEPRECATED. Apply convolution to input then add a bias.
}

