% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.cosh}
\alias{mx.nd.cosh}
\title{Returns the hyperbolic cosine  of the input array, computed element-wise.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
.. math::
   cosh(x) = 0.5\times(exp(x) + exp(-x))
}
\details{
The storage type of ``cosh`` output is always dense



Defined in src/operator/tensor/elemwise_unary_op_trig.cc:L216
}

