% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.cumsum}
\alias{mx.nd.cumsum}
\title{Return the cumulative sum of the elements along a given axis.}
\arguments{
\item{a}{NDArray-or-Symbol
Input ndarray}

\item{axis}{int or None, optional, default='None'
Axis along which the cumulative sum is computed. The default (None) is to compute the cumsum over the flattened array.}

\item{dtype}{{None, 'float16', 'float32', 'float64', 'int32', 'int64', 'int8'},optional, default='None'
Type of the returned array and of the accumulator in which the elements are summed. If dtype is not specified, it defaults to the dtype of a, unless a has an integer dtype with a precision less than that of the default platform integer. In that case, the default platform integer is used.}
}
\value{
out The result mx.ndarray
}
\description{
Defined in src/operator/numpy/np_cumsum.cc:L70
}
