% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.digamma}
\alias{mx.nd.digamma}
\title{Returns element-wise log derivative of the gamma function \
of the input.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
The storage type of ``digamma`` output is always dense
}
