% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.elemwise.add}
\alias{mx.nd.elemwise.add}
\title{Adds arguments element-wise.}
\arguments{
\item{lhs}{NDArray-or-Symbol
first input}

\item{rhs}{NDArray-or-Symbol
second input}
}
\value{
out The result mx.ndarray
}
\description{
The storage type of ``elemwise_add`` output depends on storage types of inputs
}
\details{
- elemwise_add(row_sparse, row_sparse) = row_sparse
   - elemwise_add(csr, csr) = csr
   - elemwise_add(default, csr) = default
   - elemwise_add(csr, default) = default
   - elemwise_add(default, rsp) = default
   - elemwise_add(rsp, default) = default
   - otherwise, ``elemwise_add`` generates output with default storage
}

