% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.erfinv}
\alias{mx.nd.erfinv}
\title{Returns element-wise inverse gauss error function of the input.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
Example::
}
\details{
erfinv([0, 0.5., -1.]) = [0., 0.4769, -inf]



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L909
}
