% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.expm1}
\alias{mx.nd.expm1}
\title{Returns ``exp(x) - 1`` computed element-wise on the input.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
This function provides greater precision than ``exp(x) - 1`` for small values of ``x``.
}
\details{
The storage type of ``expm1`` output depends upon the input storage type:

- expm1(default) = default
   - expm1(row_sparse) = row_sparse
   - expm1(csr) = csr



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L1000
}

