% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.gamma}
\alias{mx.nd.gamma}
\title{Returns the gamma function (extension of the factorial function \
to the reals), computed element-wise on the input array.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
The storage type of ``gamma`` output is always dense
}

