% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.gammaln}
\alias{mx.nd.gammaln}
\title{Returns element-wise log of the absolute value of the gamma function \
of the input.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
The storage type of ``gammaln`` output is always dense
}

