% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.GridGenerator}
\alias{mx.nd.GridGenerator}
\title{Generates 2D sampling grid for bilinear sampling.}
\arguments{
\item{data}{NDArray-or-Symbol
Input data to the function.}

\item{transform.type}{{'affine', 'warp'}, required
The type of transformation. For `affine`, input data should be an affine matrix of size (batch, 6). For `warp`, input data should be an optical flow of size (batch, 2, h, w).}

\item{target.shape}{Shape(tuple), optional, default=[0,0]
Specifies the output shape (H, W). This is required if transformation type is `affine`. If transformation type is `warp`, this parameter is ignored.}
}
\value{
out The result mx.ndarray
}
\description{
Generates 2D sampling grid for bilinear sampling.
}

