% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.GroupNorm}
\alias{mx.nd.GroupNorm}
\title{Group normalization.}
\arguments{
\item{data}{NDArray-or-Symbol
Input data}

\item{gamma}{NDArray-or-Symbol
gamma array}

\item{beta}{NDArray-or-Symbol
beta array}

\item{num.groups}{int, optional, default='1'
Total number of groups.}

\item{eps}{float, optional, default=9.99999975e-06
An `epsilon` parameter to prevent division by 0.}

\item{output.mean.var}{boolean, optional, default=0
Output the mean and std calculated along the given axis.}
}
\value{
out The result mx.ndarray
}
\description{
The input channels are separated into ``num_groups`` groups, each containing ``num_channels / num_groups`` channels.
The mean and standard-deviation are calculated separately over the each group.
}
\details{
.. math::

  data = data.reshape((N, num_groups, C // num_groups, ...))
  out = \frac{data - mean(data, axis)}{\sqrt{var(data, axis) + \epsilon}} * gamma + beta

Both ``gamma`` and ``beta`` are learnable parameters.



Defined in src/operator/nn/group_norm.cc:L77
}
