% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.lamb.update.phase2}
\alias{mx.nd.lamb.update.phase2}
\title{Phase II of lamb update it performs the following operations and updates grad.}
\arguments{
\item{weight}{NDArray-or-Symbol
Weight}

\item{g}{NDArray-or-Symbol
Output of lamb_update_phase 1}

\item{r1}{NDArray-or-Symbol
r1}

\item{r2}{NDArray-or-Symbol
r2}

\item{lr}{float, required
Learning rate}

\item{lower.bound}{float, optional, default=-1
Lower limit of norm of weight. If lower_bound <= 0, Lower limit is not set}

\item{upper.bound}{float, optional, default=-1
Upper limit of norm of weight. If upper_bound <= 0, Upper limit is not set}
}
\value{
out The result mx.ndarray
}
\description{
Link to paper: https://arxiv.org/pdf/1904.00962.pdf
}
\details{
.. math::
    \begin{gather*}
    if (lower_bound >= 0)
    then
         r1 = max(r1, lower_bound)
    if (upper_bound >= 0)
    then
         r1 = max(r1, upper_bound)

    if (r1 == 0 or r2 == 0)
    then
         lr = lr
    else
         lr = lr * (r1/r2)
    weight = weight - lr * g
    \end{gather*}



Defined in src/operator/optimizer_op.cc:L983
}
