% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.log10}
\alias{mx.nd.log10}
\title{Returns element-wise Base-10 logarithmic value of the input.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
``10**log10(x) = x``
}
\details{
The storage type of ``log10`` output is always dense



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L945
}

