% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.log1p}
\alias{mx.nd.log1p}
\title{Returns element-wise ``log(1 + x)`` value of the input.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
This function is more accurate than ``log(1 + x)``  for small ``x`` so that
:math:`1+x\approx 1`
}
\details{
The storage type of ``log1p`` output depends upon the input storage type:

- log1p(default) = default
   - log1p(row_sparse) = row_sparse
   - log1p(csr) = csr



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L982
}

