% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.logical.not}
\alias{mx.nd.logical.not}
\title{Returns the result of logical NOT (!) function}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
Example:
  logical_not([-2., 0., 1.]) = [0., 1., 0.]
}

