% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.max.axis}
\alias{mx.nd.max.axis}
\title{Computes the max of array elements over given axes.}
\arguments{
\item{data}{NDArray-or-Symbol
The input}

\item{axis}{Shape or None, optional, default=None
    The axis or axes along which to perform the reduction.

      The default, `axis=()`, will compute over all elements into a
      scalar array with shape `(1,)`.

      If `axis` is int, a reduction is performed on a particular axis.

      If `axis` is a tuple of ints, a reduction is performed on all the axes
      specified in the tuple.

      If `exclude` is true, reduction will be performed on the axes that are
      NOT in axis instead.

      Negative values means indexing from right to left.}

\item{keepdims}{boolean, optional, default=0
If this is set to `True`, the reduced axes are left in the result as dimension with size one.}

\item{exclude}{boolean, optional, default=0
Whether to perform reduction on axis that are NOT in axis instead.}
}
\value{
out The result mx.ndarray
}
\description{
Defined in src/operator/tensor/broadcast_reduce_op_value.cc:L191
}

