% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.mp.sgd.update}
\alias{mx.nd.mp.sgd.update}
\title{Updater function for multi-precision sgd optimizer}
\arguments{
\item{weight}{NDArray-or-Symbol
Weight}

\item{grad}{NDArray-or-Symbol
gradient}

\item{weight32}{NDArray-or-Symbol
Weight32}

\item{lr}{float, required
Learning rate}

\item{wd}{float, optional, default=0
Weight decay augments the objective function with a regularization term that penalizes large weights. The penalty scales with the square of the magnitude of each weight.}

\item{rescale.grad}{float, optional, default=1
Rescale gradient to grad = rescale_grad*grad.}

\item{clip.gradient}{float, optional, default=-1
Clip gradient to the range of [-clip_gradient, clip_gradient] If clip_gradient <= 0, gradient clipping is turned off. grad = max(min(grad, clip_gradient), -clip_gradient).}

\item{lazy.update}{boolean, optional, default=1
If true, lazy updates are applied if gradient's stype is row_sparse.}
}
\value{
out The result mx.ndarray
}
\description{
Updater function for multi-precision sgd optimizer
}

