% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.multi.all.finite}
\alias{mx.nd.multi.all.finite}
\title{Check if all the float numbers in all the arrays are finite (used for AMP)}
\arguments{
\item{data}{NDArray-or-Symbol[]
Arrays}

\item{num.arrays}{int, optional, default='1'
Number of arrays.}

\item{init.output}{boolean, optional, default=1
Initialize output to 1.}
}
\value{
out The result mx.ndarray
}
\description{
Defined in src/operator/contrib/all_finite.cc:L133
}
