% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.multi.sgd.update}
\alias{mx.nd.multi.sgd.update}
\title{Update function for Stochastic Gradient Descent (SDG) optimizer.}
\arguments{
\item{data}{NDArray-or-Symbol[]
Weights}

\item{lrs}{tuple of <float>, required
Learning rates.}

\item{wds}{tuple of <float>, required
Weight decay augments the objective function with a regularization term that penalizes large weights. The penalty scales with the square of the magnitude of each weight.}

\item{rescale.grad}{float, optional, default=1
Rescale gradient to grad = rescale_grad*grad.}

\item{clip.gradient}{float, optional, default=-1
Clip gradient to the range of [-clip_gradient, clip_gradient] If clip_gradient <= 0, gradient clipping is turned off. grad = max(min(grad, clip_gradient), -clip_gradient).}

\item{num.weights}{int, optional, default='1'
Number of updated weights.}
}
\value{
out The result mx.ndarray
}
\description{
It updates the weights using::
}
\details{
weight = weight - learning_rate * (gradient + wd * weight)



Defined in src/operator/optimizer_op.cc:L329
}
