% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.multi.sum.sq}
\alias{mx.nd.multi.sum.sq}
\title{Compute the sums of squares of multiple arrays}
\arguments{
\item{data}{NDArray-or-Symbol[]
Arrays}

\item{num.arrays}{int, required
number of input arrays.}
}
\value{
out The result mx.ndarray
}
\description{
Defined in src/operator/contrib/multi_sum_sq.cc:L36
}
