% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.negative}
\alias{mx.nd.negative}
\title{Numerical negative of the argument, element-wise.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
The storage type of ``negative`` output depends upon the input storage type:
}
\details{
- negative(default) = default
   - negative(row_sparse) = row_sparse
   - negative(csr) = csr
}

