% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.random.exponential}
\alias{mx.nd.random.exponential}
\title{Draw random samples from an exponential distribution.}
\arguments{
\item{lam}{float, optional, default=1
Lambda parameter (rate) of the exponential distribution.}

\item{shape}{Shape(tuple), optional, default=[]
Shape of the output.}

\item{ctx}{string, optional, default=''
Context of output, in format [cpu|gpu|cpu_pinned](n). Only used for imperative calls.}

\item{dtype}{{'None', 'float16', 'float32', 'float64'},optional, default='None'
DType of the output in case this can't be inferred. Defaults to float32 if not defined (dtype=None).}
}
\value{
out The result mx.ndarray
}
\description{
Samples are distributed according to an exponential distribution parametrized by *lambda* (rate).
}
\details{
Example::

exponential(lam=4, shape=(2,2)) = [[ 0.0097189 ,  0.08999364],
                                      [ 0.04146638,  0.31715935]]

Defined in src/operator/random/sample_op.cc:L136
}

