% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.random.gamma}
\alias{mx.nd.random.gamma}
\title{Draw random samples from a gamma distribution.}
\arguments{
\item{alpha}{float, optional, default=1
Alpha parameter (shape) of the gamma distribution.}

\item{beta}{float, optional, default=1
Beta parameter (scale) of the gamma distribution.}

\item{shape}{Shape(tuple), optional, default=[]
Shape of the output.}

\item{ctx}{string, optional, default=''
Context of output, in format [cpu|gpu|cpu_pinned](n). Only used for imperative calls.}

\item{dtype}{{'None', 'float16', 'float32', 'float64'},optional, default='None'
DType of the output in case this can't be inferred. Defaults to float32 if not defined (dtype=None).}
}
\value{
out The result mx.ndarray
}
\description{
Samples are distributed according to a gamma distribution parametrized by *alpha* (shape) and *beta* (scale).
}
\details{
Example::

gamma(alpha=9, beta=0.5, shape=(2,2)) = [[ 7.10486984,  3.37695289],
                                            [ 3.91697288,  3.65933681]]

Defined in src/operator/random/sample_op.cc:L124
}

