% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.random.poisson}
\alias{mx.nd.random.poisson}
\title{Draw random samples from a Poisson distribution.}
\arguments{
\item{lam}{float, optional, default=1
Lambda parameter (rate) of the Poisson distribution.}

\item{shape}{Shape(tuple), optional, default=[]
Shape of the output.}

\item{ctx}{string, optional, default=''
Context of output, in format [cpu|gpu|cpu_pinned](n). Only used for imperative calls.}

\item{dtype}{{'None', 'float16', 'float32', 'float64'},optional, default='None'
DType of the output in case this can't be inferred. Defaults to float32 if not defined (dtype=None).}
}
\value{
out The result mx.ndarray
}
\description{
Samples are distributed according to a Poisson distribution parametrized by *lambda* (rate).
Samples will always be returned as a floating point data type.
}
\details{
Example::

poisson(lam=4, shape=(2,2)) = [[ 5.,  2.],
                                  [ 4.,  6.]]

Defined in src/operator/random/sample_op.cc:L149
}

