% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.random.randint}
\alias{mx.nd.random.randint}
\title{Draw random samples from a discrete uniform distribution.}
\arguments{
\item{low}{long, required
Lower bound of the distribution.}

\item{high}{long, required
Upper bound of the distribution.}

\item{shape}{Shape(tuple), optional, default=None
Shape of the output.}

\item{ctx}{string, optional, default=''
Context of output, in format [cpu|gpu|cpu_pinned](n). Only used for imperative calls.}

\item{dtype}{{'None', 'int32', 'int64'},optional, default='None'
DType of the output in case this can't be inferred. Defaults to int32 if not defined (dtype=None).}
}
\value{
out The result mx.ndarray
}
\description{
Samples are uniformly distributed over the half-open interval *[low, high)*
(includes *low*, but excludes *high*).
}
\details{
Example::

   randint(low=0, high=5, shape=(2,2)) = [[ 0,  2],
                                          [ 3,  1]]



Defined in src/operator/random/sample_op.cc:L194
}
