% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.ravel.multi.index}
\alias{mx.nd.ravel.multi.index}
\title{Converts a batch of index arrays into an array of flat indices. The operator follows numpy conventions so a single multi index is given by a column of the input matrix.}
\arguments{
\item{data}{NDArray-or-Symbol
Batch of multi-indices}

\item{shape}{Shape(tuple), optional, default=[]
Shape of the array into which the multi-indices apply.}
}
\value{
out The result mx.ndarray
}
\description{
Examples::
   
   A = [[3,6,6],[4,5,1]]
   ravel(A, shape=(7,6)) = [22,41,37]
}
\details{
Defined in src/operator/tensor/ravel.cc:L41
}

