% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.rcbrt}
\alias{mx.nd.rcbrt}
\title{Returns element-wise inverse cube-root value of the input.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
.. math::
   rcbrt(x) = 1/\sqrt[3]{x}
}
\details{
Example::

rcbrt([1,8,-125]) = [1.0, 0.5, -0.2]



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L898
}

