% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.reset.arrays}
\alias{mx.nd.reset.arrays}
\title{Set to zero multiple arrays}
\arguments{
\item{data}{NDArray-or-Symbol[]
Arrays}

\item{num.arrays}{int, required
number of input arrays.}
}
\value{
out The result mx.ndarray
}
\description{
Defined in src/operator/contrib/reset_arrays.cc:L36
}
