% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.rsqrt}
\alias{mx.nd.rsqrt}
\title{Returns element-wise inverse square-root value of the input.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
.. math::
   rsqrt(x) = 1/\sqrt{x}
}
\details{
Example::

rsqrt([4,9,16]) = [0.5, 0.33333334, 0.25]

The storage type of ``rsqrt`` output is always dense



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L858
}

