% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.shape.array}
\alias{mx.nd.shape.array}
\title{Returns a 1D int64 array containing the shape of data.}
\arguments{
\item{data}{NDArray-or-Symbol
Input Array.}
}
\value{
out The result mx.ndarray
}
\description{
Example::
}
\details{
shape_array([[1,2,3,4], [5,6,7,8]]) = [2,4]



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L574
}
