% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.shuffle}
\alias{mx.nd.shuffle}
\title{Randomly shuffle the elements.}
\arguments{
\item{data}{NDArray-or-Symbol
Data to be shuffled.}
}
\value{
out The result mx.ndarray
}
\description{
This shuffles the array along the first axis.
The order of the elements in each subarray does not change.
For example, if a 2D array is given, the order of the rows randomly changes,
but the order of the elements in each row does not change.
}

