% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.sigmoid}
\alias{mx.nd.sigmoid}
\title{Computes sigmoid of x element-wise.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
.. math::
   y = 1 / (1 + exp(-x))
}
\details{
The storage type of ``sigmoid`` output is always dense



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L101
}

