% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.sinh}
\alias{mx.nd.sinh}
\title{Returns the hyperbolic sine of the input array, computed element-wise.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
.. math::
   sinh(x) = 0.5\times(exp(x) - exp(-x))
}
\details{
The storage type of ``sinh`` output depends upon the input storage type:

- sinh(default) = default
   - sinh(row_sparse) = row_sparse
   - sinh(csr) = csr



Defined in src/operator/tensor/elemwise_unary_op_trig.cc:L201
}

