% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.softsign}
\alias{mx.nd.softsign}
\title{Computes softsign of x element-wise.}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}
}
\value{
out The result mx.ndarray
}
\description{
.. math::
   y = x / (1 + abs(x))
}
\details{
The storage type of ``softsign`` output is always dense



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L145
}

