% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.SVMOutput}
\alias{mx.nd.SVMOutput}
\title{Computes support vector machine based transformation of the input.}
\arguments{
\item{data}{NDArray-or-Symbol
Input data for SVM transformation.}

\item{label}{NDArray-or-Symbol
Class label for the input data.}

\item{margin}{float, optional, default=1
The loss function penalizes outputs that lie outside this margin. Default margin is 1.}

\item{regularization.coefficient}{float, optional, default=1
Regularization parameter for the SVM. This balances the tradeoff between coefficient size and error.}

\item{use.linear}{boolean, optional, default=0
Whether to use L1-SVM objective. L2-SVM objective is used by default.}
}
\value{
out The result mx.ndarray
}
\description{
This tutorial demonstrates using SVM as output layer for classification instead of softmax:
https://github.com/dmlc/mxnet/tree/master/example/svm_mnist.
}

