% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.swapaxes}
\alias{mx.nd.swapaxes}
\title{Interchanges two axes of an array.}
\arguments{
\item{data}{NDArray-or-Symbol
Input array.}

\item{dim1}{int (non-negative), optional, default=0
the first axis to be swapped.}

\item{dim2}{int (non-negative), optional, default=0
the second axis to be swapped.}
}
\value{
out The result mx.ndarray
}
\description{
Examples::
}
\details{
x = [[1, 2, 3]])
  swapaxes(x, 0, 1) = [[ 1],
                       [ 2],
                       [ 3]]

x = [[[ 0, 1],
        [ 2, 3]],
       [[ 4, 5],
        [ 6, 7]]]  // (2,2,2) array

swapaxes(x, 0, 2) = [[[ 0, 4],
                       [ 2, 6]],
                      [[ 1, 5],
                       [ 3, 7]]]

Defined in src/operator/swapaxis.cc:L70
}

