% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizer.R
\name{mx.opt.sgd}
\alias{mx.opt.sgd}
\title{Create an SGD optimizer with respective parameters.
Perform SGD with momentum update}
\usage{
mx.opt.sgd(learning.rate = 0.01, momentum = 0, wd = 0, rescale.grad = 1,
  clip_gradient = -1, lr_scheduler = NULL)
}
\arguments{
\item{learning.rate}{float, default=0.01
The initial learning rate.}

\item{momentum}{float, default=0
The momentumvalue}

\item{wd}{float, default=0.0
L2 regularization coefficient add to all the weights.}

\item{rescale.grad}{float, default=1.0
rescaling factor of gradient.}

\item{clip_gradient}{float, optional, default=-1 (no clipping if < 0)
clip gradient in range [-clip_gradient, clip_gradient].}

\item{lr_scheduler}{function, optional
The learning rate scheduler.}
}
\description{
Create an SGD optimizer with respective parameters.
Perform SGD with momentum update
}

