% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{mx.profiler.state}
\alias{mx.profiler.state}
\title{Set up the profiler state to record operator.}
\usage{
mx.profiler.state(state = MX.PROF.STATE$STOP)
}
\arguments{
\item{state}{Indicting whether to run the profiler, can be 'MX.PROF.STATE$RUN' or 'MX.PROF.STATE$STOP'. Default is `MX.PROF.STATE$STOP`.}

\item{filename}{The name of output trace file. Default is 'profile.json'}
}
\description{
Set up the profiler state to record operator.
}

