% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mx.rnorm}
\alias{mx.rnorm}
\title{Generate nomal distribution with mean and sd.}
\usage{
mx.rnorm(shape, mean = 0, sd = 1, ctx = NULL)
}
\arguments{
\item{shape}{Dimension, The shape(dimension) of the result.}

\item{mean}{numeric, The mean of distribution.}

\item{sd}{numeric, The standard deviations.}

\item{ctx, }{optional The context device of the array. mx.ctx.default() will be used in default.}
}
\description{
Generate nomal distribution with mean and sd.
}
\examples{

mx.set.seed(0)
as.array(mx.runif(2))
# 0.5488135 0.5928446
mx.set.seed(0)
as.array(mx.rnorm(2))
# 2.212206 1.163079

}

