% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mx.serialize}
\alias{mx.serialize}
\title{Serialize MXNet model into RData-compatiable format.}
\usage{
mx.serialize(model)
}
\arguments{
\item{model}{The mxnet model}
}
\description{
Serialize MXNet model into RData-compatiable format.
}

