% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.all_finite}
\alias{mx.symbol.all_finite}
\title{all_finite:Check if all the float numbers in the array are finite (used for AMP)}
\usage{
mx.symbol.all_finite(...)
}
\arguments{
\item{data}{NDArray
Array}

\item{init.output}{boolean, optional, default=1
Initialize output to 1.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Defined in src/operator/contrib/all_finite.cc:L101
}
