% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.amp_cast}
\alias{mx.symbol.amp_cast}
\title{amp_cast:Cast function between low precision float/FP32 used by AMP.}
\usage{
mx.symbol.amp_cast(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input.}

\item{dtype}{{'float16', 'float32', 'float64', 'int32', 'int64', 'int8', 'uint8'}, required
Output data type.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
It casts only between low precision float/FP32 and does not do anything for other types.
}
\details{
Defined in src/operator/tensor/amp_cast.cc:L37
}
