% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.amp_multicast}
\alias{mx.symbol.amp_multicast}
\title{amp_multicast:Cast function used by AMP, that casts its inputs to the common widest type.}
\usage{
mx.symbol.amp_multicast(...)
}
\arguments{
\item{data}{NDArray-or-Symbol[]
Weights}

\item{num.outputs}{int, required
Number of input/output pairs to be casted to the widest type.}

\item{cast.narrow}{boolean, optional, default=0
Whether to cast to the narrowest type}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
It casts only between low precision float/FP32 and does not do anything for other types.
}
\details{
Defined in src/operator/tensor/amp_cast.cc:L71
}
