% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.arccos}
\alias{mx.symbol.arccos}
\title{arccos:Returns element-wise inverse cosine of the input array.}
\usage{
mx.symbol.arccos(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
The input should be in range `[-1, 1]`.
The output is in the closed interval :math:`[0, \pi]`
}
\details{
.. math::
   arccos([-1, -.707, 0, .707, 1]) = [\pi, 3\pi/4, \pi/2, \pi/4, 0]

The storage type of ``arccos`` output is always dense



Defined in src/operator/tensor/elemwise_unary_op_trig.cc:L123
}

