% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.argmax}
\alias{mx.symbol.argmax}
\title{argmax:Returns indices of the maximum values along an axis.}
\usage{
mx.symbol.argmax(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input}

\item{axis}{int or None, optional, default='None'
The axis along which to perform the reduction. Negative values means indexing from right to left. ``Requires axis to be set as int, because global reduction is not supported yet.``}

\item{keepdims}{boolean, optional, default=0
If this is set to `True`, the reduced axis is left in the result as dimension with size one.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
In the case of multiple occurrences of maximum values, the indices corresponding to the first occurrence
are returned.
}
\details{
Examples::

x = [[ 0.,  1.,  2.],
       [ 3.,  4.,  5.]]

// argmax along axis 0
  argmax(x, axis=0) = [ 1.,  1.,  1.]

// argmax along axis 1
  argmax(x, axis=1) = [ 2.,  2.]

// argmax along axis 1 keeping same dims as an input array
  argmax(x, axis=1, keepdims=True) = [[ 2.],
                                      [ 2.]]



Defined in src/operator/tensor/broadcast_reduce_op_index.cc:L52
}

