% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.batch_dot}
\alias{mx.symbol.batch_dot}
\title{batch_dot:Batchwise dot product.}
\usage{
mx.symbol.batch_dot(...)
}
\arguments{
\item{lhs}{NDArray-or-Symbol
The first input}

\item{rhs}{NDArray-or-Symbol
The second input}

\item{transpose.a}{boolean, optional, default=0
If true then transpose the first input before dot.}

\item{transpose.b}{boolean, optional, default=0
If true then transpose the second input before dot.}

\item{forward.stype}{{None, 'csr', 'default', 'row_sparse'},optional, default='None'
The desired storage type of the forward output given by user, if thecombination of input storage types and this hint does not matchany implemented ones, the dot operator will perform fallback operationand still produce an output of the desired storage type.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
``batch_dot`` is used to compute dot product of ``x`` and ``y`` when ``x`` and
``y`` are data in batch, namely N-D (N >= 3) arrays in shape of `(B0, ..., B_i, :, :)`.
}
\details{
For example, given ``x`` with shape `(B_0, ..., B_i, N, M)` and ``y`` with shape
`(B_0, ..., B_i, M, K)`, the result array will have shape `(B_0, ..., B_i, N, K)`,
which is computed by::

   batch_dot(x,y)[b_0, ..., b_i, :, :] = dot(x[b_0, ..., b_i, :, :], y[b_0, ..., b_i, :, :])



Defined in src/operator/tensor/dot.cc:L127
}
