% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.batch_take}
\alias{mx.symbol.batch_take}
\title{batch_take:Takes elements from a data batch.}
\usage{
mx.symbol.batch_take(...)
}
\arguments{
\item{a}{NDArray-or-Symbol
The input array}

\item{indices}{NDArray-or-Symbol
The index array}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
.. note::
  `batch_take` is deprecated. Use `pick` instead.
}
\details{
Given an input array of shape ``(d0, d1)`` and indices of shape ``(i0,)``, the result will be
an output array of shape ``(i0,)`` with::

output[i] = input[i, indices[i]]

Examples::

x = [[ 1.,  2.],
       [ 3.,  4.],
       [ 5.,  6.]]

// takes elements with specified indices
  batch_take(x, [0,1,0]) = [ 1.  4.  5.]



Defined in src/operator/tensor/indexing_op.cc:L750
}

