% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.broadcast_logical_or}
\alias{mx.symbol.broadcast_logical_or}
\title{broadcast_logical_or:Returns the result of element-wise **logical or** with broadcasting.}
\usage{
mx.symbol.broadcast_logical_or(...)
}
\arguments{
\item{lhs}{NDArray-or-Symbol
First input to the function}

\item{rhs}{NDArray-or-Symbol
Second input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Example::
}
\details{
x = [[ 1.,  1.,  0.],
        [ 1.,  1.,  0.]]

y = [[ 1.],
        [ 0.]]

broadcast_logical_or(x, y) = [[ 1.,  1.,  1.],
                                 [ 1.,  1.,  0.]]



Defined in src/operator/tensor/elemwise_binary_broadcast_op_logic.cc:L172
}

