% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.broadcast_mod}
\alias{mx.symbol.broadcast_mod}
\title{broadcast_mod:Returns element-wise modulo of the input arrays with broadcasting.}
\usage{
mx.symbol.broadcast_mod(...)
}
\arguments{
\item{lhs}{NDArray-or-Symbol
First input to the function}

\item{rhs}{NDArray-or-Symbol
Second input to the function}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Example::
}
\details{
x = [[ 8.,  8.,  8.],
        [ 8.,  8.,  8.]]

y = [[ 2.],
        [ 3.]]

broadcast_mod(x, y) = [[ 0.,  0.,  0.],
                          [ 2.,  2.,  2.]]



Defined in src/operator/tensor/elemwise_binary_broadcast_op_basic.cc:L222
}

