% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.cast}
\alias{mx.symbol.cast}
\title{cast:Casts all elements of the input to a new type.}
\usage{
mx.symbol.cast(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input.}

\item{dtype}{{'float16', 'float32', 'float64', 'int32', 'int64', 'int8', 'uint8'}, required
Output data type.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
.. note:: ``Cast`` is deprecated. Use ``cast`` instead.
}
\details{
Example::

cast([0.9, 1.3], dtype='int32') = [0, 1]
   cast([1e20, 11.1], dtype='float16') = [inf, 11.09375]
   cast([300, 11.1, 10.9, -1, -3], dtype='uint8') = [44, 11, 10, 255, 253]



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L594
}

