% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.digamma}
\alias{mx.symbol.digamma}
\title{digamma:Returns element-wise log derivative of the gamma function \
of the input.}
\usage{
mx.symbol.digamma(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
The storage type of ``digamma`` output is always dense
}
