% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.elemwise_div}
\alias{mx.symbol.elemwise_div}
\title{elemwise_div:Divides arguments element-wise.}
\usage{
mx.symbol.elemwise_div(...)
}
\arguments{
\item{lhs}{NDArray-or-Symbol
first input}

\item{rhs}{NDArray-or-Symbol
second input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
The storage type of ``elemwise_div`` output is always dense
}

