% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.elemwise_sub}
\alias{mx.symbol.elemwise_sub}
\title{elemwise_sub:Subtracts arguments element-wise.}
\usage{
mx.symbol.elemwise_sub(...)
}
\arguments{
\item{lhs}{NDArray-or-Symbol
first input}

\item{rhs}{NDArray-or-Symbol
second input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
The storage type of ``elemwise_sub`` output depends on storage types of inputs
}
\details{
- elemwise_sub(row_sparse, row_sparse) = row_sparse
   - elemwise_sub(csr, csr) = csr
   - elemwise_sub(default, csr) = default
   - elemwise_sub(csr, default) = default
   - elemwise_sub(default, rsp) = default
   - elemwise_sub(rsp, default) = default
   - otherwise, ``elemwise_sub`` generates output with default storage
}

