% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.erf}
\alias{mx.symbol.erf}
\title{erf:Returns element-wise gauss error function of the input.}
\usage{
mx.symbol.erf(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Example::
}
\details{
erf([0, -1., 10.]) = [0., -0.8427, 1.]



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L886
}
