% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.erfinv}
\alias{mx.symbol.erfinv}
\title{erfinv:Returns element-wise inverse gauss error function of the input.}
\usage{
mx.symbol.erfinv(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Example::
}
\details{
erfinv([0, 0.5., -1.]) = [0., 0.4769, -inf]



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L907
}
