% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.expand_dims}
\alias{mx.symbol.expand_dims}
\title{expand_dims:Inserts a new axis of size 1 into the array shape}
\usage{
mx.symbol.expand_dims(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
Source input}

\item{axis}{int, required
Position where new axis is to be inserted. Suppose that the input `NDArray`'s dimension is `ndim`, the range of the inserted axis is `[-ndim, ndim]`}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
For example, given ``x`` with shape ``(2,3,4)``, then ``expand_dims(x, axis=1)``
will return a new array with shape ``(2,1,3,4)``.
}
\details{
Defined in src/operator/tensor/matrix_op.cc:L348
}

