% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.flip}
\alias{mx.symbol.flip}
\title{flip:Reverses the order of elements along given axis while preserving array shape.}
\usage{
mx.symbol.flip(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
Input data array}

\item{axis}{Shape(tuple), required
The axis which to reverse elements.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Note: reverse and flip are equivalent. We use reverse in the following examples.
}
\details{
Examples::

x = [[ 0.,  1.,  2.,  3.,  4.],
       [ 5.,  6.,  7.,  8.,  9.]]

reverse(x, axis=0) = [[ 5.,  6.,  7.,  8.,  9.],
                        [ 0.,  1.,  2.,  3.,  4.]]

reverse(x, axis=1) = [[ 4.,  3.,  2.,  1.,  0.],
                        [ 9.,  8.,  7.,  6.,  5.]]

Defined in src/operator/tensor/matrix_op.cc:L794
}

