% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.gammaln}
\alias{mx.symbol.gammaln}
\title{gammaln:Returns element-wise log of the absolute value of the gamma function \
of the input.}
\usage{
mx.symbol.gammaln(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
The storage type of ``gammaln`` output is always dense
}

