% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.linalg_inverse}
\alias{mx.symbol.linalg_inverse}
\title{linalg_inverse:Compute the inverse of a matrix.
Input is a tensor *A* of dimension *n >= 2*.}
\usage{
mx.symbol.linalg_inverse(...)
}
\arguments{
\item{A}{NDArray-or-Symbol
Tensor of square matrix}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
If *n=2*, *A* is a square matrix. We compute:
}
\details{
*out* = *A*\ :sup:`-1`

If *n>2*, *inverse* is performed separately on the trailing two dimensions
for all inputs (batch mode).

.. note:: The operator supports float32 and float64 data types only.

Examples::

   Single matrix inverse
   A = [[1., 4.], [2., 3.]]
   inverse(A) = [[-0.6, 0.8], [0.4, -0.2]]

   Batch matrix inverse
   A = [[[1., 4.], [2., 3.]],
        [[1., 3.], [2., 4.]]]
   inverse(A) = [[[-0.6, 0.8], [0.4, -0.2]],
                 [[-2., 1.5], [1., -0.5]]]


Defined in src/operator/tensor/la_op.cc:L920
}
