% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.linalg_trmm}
\alias{mx.symbol.linalg_trmm}
\title{linalg_trmm:Performs multiplication with a lower triangular matrix.
Input are tensors *A*, *B*, each of dimension *n >= 2* and having the same shape
on the leading *n-2* dimensions.}
\usage{
mx.symbol.linalg_trmm(...)
}
\arguments{
\item{A}{NDArray-or-Symbol
Tensor of lower triangular matrices}

\item{B}{NDArray-or-Symbol
Tensor of matrices}

\item{transpose}{boolean, optional, default=0
Use transposed of the triangular matrix}

\item{rightside}{boolean, optional, default=0
Multiply triangular matrix from the right to non-triangular one.}

\item{alpha}{double, optional, default=1
Scalar factor to be applied to the result.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
If *n=2*, *A* must be lower triangular. The operator performs the BLAS3 function
*trmm*:
}
\details{
*out* = *alpha* \* *op*\ (*A*) \* *B*

if *rightside=False*, or

*out* = *alpha* \* *B* \* *op*\ (*A*)

if *rightside=True*. Here, *alpha* is a scalar parameter, and *op()* is either the
identity or the matrix transposition (depending on *transpose*).

If *n>2*, *trmm* is performed separately on the trailing two dimensions for all inputs
(batch mode).

.. note:: The operator supports float32 and float64 data types only.

Examples::

// Single triangular matrix multiply
   A = [[1.0, 0], [1.0, 1.0]]
   B = [[1.0, 1.0, 1.0], [1.0, 1.0, 1.0]]
   trmm(A, B, alpha=2.0) = [[2.0, 2.0, 2.0], [4.0, 4.0, 4.0]]

// Batch triangular matrix multiply
   A = [[[1.0, 0], [1.0, 1.0]], [[1.0, 0], [1.0, 1.0]]]
   B = [[[1.0, 1.0, 1.0], [1.0, 1.0, 1.0]], [[0.5, 0.5, 0.5], [0.5, 0.5, 0.5]]]
   trmm(A, B, alpha=2.0) = [[[2.0, 2.0, 2.0], [4.0, 4.0, 4.0]],
                            [[1.0, 1.0, 1.0], [2.0, 2.0, 2.0]]]

Defined in src/operator/tensor/la_op.cc:L316
}

