% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.log2}
\alias{mx.symbol.log2}
\title{log2:Returns element-wise Base-2 logarithmic value of the input.}
\usage{
mx.symbol.log2(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
``2**log2(x) = x``
}
\details{
The storage type of ``log2`` output is always dense



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L957
}

