% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.mp_lamb_update_phase2}
\alias{mx.symbol.mp_lamb_update_phase2}
\title{mp_lamb_update_phase2:Mixed Precision version Phase II of lamb update 
it performs the following operations and updates grad.}
\usage{
mx.symbol.mp_lamb_update_phase2(...)
}
\arguments{
\item{weight}{NDArray-or-Symbol
Weight}

\item{g}{NDArray-or-Symbol
Output of mp_lamb_update_phase 1}

\item{r1}{NDArray-or-Symbol
r1}

\item{r2}{NDArray-or-Symbol
r2}

\item{weight32}{NDArray-or-Symbol
Weight32}

\item{lr}{float, required
Learning rate}

\item{lower.bound}{float, optional, default=-1
Lower limit of norm of weight. If lower_bound <= 0, Lower limit is not set}

\item{upper.bound}{float, optional, default=-1
Upper limit of norm of weight. If upper_bound <= 0, Upper limit is not set}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Link to paper: https://arxiv.org/pdf/1904.00962.pdf
}
\details{
.. math::
              \begin{gather*}
              if (lower_bound >= 0)
              then
                   r1 = max(r1, lower_bound)
              if (upper_bound >= 0)
              then
                   r1 = max(r1, upper_bound)

              if (r1 == 0 or r2 == 0)
              then
                   lr = lr
              else
                   lr = lr * (r1/r2)
              weight32 = weight32 - lr * g
              weight(float16) = weight32
              \end{gather*}

          

Defined in src/operator/optimizer_op.cc:L1075
}
