% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.mp_nag_mom_update}
\alias{mx.symbol.mp_nag_mom_update}
\title{mp_nag_mom_update:Update function for multi-precision Nesterov Accelerated Gradient( NAG) optimizer.}
\usage{
mx.symbol.mp_nag_mom_update(...)
}
\arguments{
\item{weight}{NDArray-or-Symbol
Weight}

\item{grad}{NDArray-or-Symbol
Gradient}

\item{mom}{NDArray-or-Symbol
Momentum}

\item{weight32}{NDArray-or-Symbol
Weight32}

\item{lr}{float, required
Learning rate}

\item{momentum}{float, optional, default=0
The decay rate of momentum estimates at each epoch.}

\item{wd}{float, optional, default=0
Weight decay augments the objective function with a regularization term that penalizes large weights. The penalty scales with the square of the magnitude of each weight.}

\item{rescale.grad}{float, optional, default=1
Rescale gradient to grad = rescale_grad*grad.}

\item{clip.gradient}{float, optional, default=-1
Clip gradient to the range of [-clip_gradient, clip_gradient] If clip_gradient <= 0, gradient clipping is turned off. grad = max(min(grad, clip_gradient), -clip_gradient).}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Defined in src/operator/optimizer_op.cc:L736
}
