% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.multi_all_finite}
\alias{mx.symbol.multi_all_finite}
\title{multi_all_finite:Check if all the float numbers in all the arrays are finite (used for AMP)}
\usage{
mx.symbol.multi_all_finite(...)
}
\arguments{
\item{data}{NDArray-or-Symbol[]
Arrays}

\item{num.arrays}{int, optional, default='1'
Number of arrays.}

\item{init.output}{boolean, optional, default=1
Initialize output to 1.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Defined in src/operator/contrib/all_finite.cc:L133
}
