% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.multi_lars}
\alias{mx.symbol.multi_lars}
\title{multi_lars:Compute the LARS coefficients of multiple weights and grads from their sums of square"}
\usage{
mx.symbol.multi_lars(...)
}
\arguments{
\item{lrs}{NDArray-or-Symbol
Learning rates to scale by LARS coefficient}

\item{weights.sum.sq}{NDArray-or-Symbol
sum of square of weights arrays}

\item{grads.sum.sq}{NDArray-or-Symbol
sum of square of gradients arrays}

\item{wds}{NDArray-or-Symbol
weight decays}

\item{eta}{float, required
LARS eta}

\item{eps}{float, required
LARS eps}

\item{rescale.grad}{float, optional, default=1
Gradient rescaling factor}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Defined in src/operator/contrib/multi_lars.cc:L37
}
