% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.multi_sum_sq}
\alias{mx.symbol.multi_sum_sq}
\title{multi_sum_sq:Compute the sums of squares of multiple arrays}
\usage{
mx.symbol.multi_sum_sq(...)
}
\arguments{
\item{data}{NDArray-or-Symbol[]
Arrays}

\item{num.arrays}{int, required
number of input arrays.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Defined in src/operator/contrib/multi_sum_sq.cc:L36
}
