% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.normal}
\alias{mx.symbol.normal}
\title{normal:Draw random samples from a normal (Gaussian) distribution.}
\usage{
mx.symbol.normal(...)
}
\arguments{
\item{loc}{float, optional, default=0
Mean of the distribution.}

\item{scale}{float, optional, default=1
Standard deviation of the distribution.}

\item{shape}{Shape(tuple), optional, default=[]
Shape of the output.}

\item{ctx}{string, optional, default=''
Context of output, in format [cpu|gpu|cpu_pinned](n). Only used for imperative calls.}

\item{dtype}{{'None', 'float16', 'float32', 'float64'},optional, default='None'
DType of the output in case this can't be inferred. Defaults to float32 if not defined (dtype=None).}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
.. note:: The existing alias ``normal`` is deprecated.
}
\details{
Samples are distributed according to a normal distribution parametrized by *loc* (mean) and *scale*
(standard deviation).

Example::

normal(loc=0, scale=1, shape=(2,2)) = [[ 1.89171135, -1.16881478],
                                          [-1.23474145,  1.55807114]]

Defined in src/operator/random/sample_op.cc:L112
}

