% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.ones_like}
\alias{mx.symbol.ones_like}
\title{ones_like:Return an array of ones with the same shape and type
as the input array.}
\usage{
mx.symbol.ones_like(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Examples::
}
\details{
x = [[ 0.,  0.,  0.],
       [ 0.,  0.,  0.]]

ones_like(x) = [[ 1.,  1.,  1.],
                  [ 1.,  1.,  1.]]
}

