% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.preloaded_multi_sgd_update}
\alias{mx.symbol.preloaded_multi_sgd_update}
\title{preloaded_multi_sgd_update:Update function for Stochastic Gradient Descent (SDG) optimizer.}
\usage{
mx.symbol.preloaded_multi_sgd_update(...)
}
\arguments{
\item{data}{NDArray-or-Symbol[]
Weights, gradients, learning rates and weight decays}

\item{rescale.grad}{float, optional, default=1
Rescale gradient to grad = rescale_grad*grad.}

\item{clip.gradient}{float, optional, default=-1
Clip gradient to the range of [-clip_gradient, clip_gradient] If clip_gradient <= 0, gradient clipping is turned off. grad = max(min(grad, clip_gradient), -clip_gradient).}

\item{num.weights}{int, optional, default='1'
Number of updated weights.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
It updates the weights using::
}
\details{
weight = weight - learning_rate * (gradient + wd * weight)



Defined in src/operator/contrib/preloaded_multi_sgd.cc:L42
}
