% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.random_negative_binomial}
\alias{mx.symbol.random_negative_binomial}
\title{random_negative_binomial:Draw random samples from a negative binomial distribution.}
\usage{
mx.symbol.random_negative_binomial(...)
}
\arguments{
\item{k}{int, optional, default='1'
Limit of unsuccessful experiments.}

\item{p}{float, optional, default=1
Failure probability in each experiment.}

\item{shape}{Shape(tuple), optional, default=[]
Shape of the output.}

\item{ctx}{string, optional, default=''
Context of output, in format [cpu|gpu|cpu_pinned](n). Only used for imperative calls.}

\item{dtype}{{'None', 'float16', 'float32', 'float64'},optional, default='None'
DType of the output in case this can't be inferred. Defaults to float32 if not defined (dtype=None).}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Samples are distributed according to a negative binomial distribution parametrized by
*k* (limit of unsuccessful experiments) and *p* (failure probability in each experiment).
Samples will always be returned as a floating point data type.
}
\details{
Example::

negative_binomial(k=3, p=0.4, shape=(2,2)) = [[ 4.,  7.],
                                                 [ 2.,  5.]]

Defined in src/operator/random/sample_op.cc:L163
}

