% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.random_pdf_normal}
\alias{mx.symbol.random_pdf_normal}
\title{random_pdf_normal:Computes the value of the PDF of *sample* of
normal distributions with parameters *mu* (mean) and *sigma* (standard deviation).}
\usage{
mx.symbol.random_pdf_normal(...)
}
\arguments{
\item{sample}{NDArray-or-Symbol
Samples from the distributions.}

\item{mu}{NDArray-or-Symbol
Means of the distributions.}

\item{is.log}{boolean, optional, default=0
If set, compute the density of the log-probability instead of the probability.}

\item{sigma}{NDArray-or-Symbol
Standard deviations of the distributions.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
*mu* and *sigma* must have the same shape, which must match the leftmost subshape
of *sample*.  That is, *sample* can have the same shape as *mu* and *sigma*, in which
case the output contains one density per distribution, or *sample* can be a tensor
of tensors with that shape, in which case the output is a tensor of densities such that
the densities at index *i* in the output are given by the samples at index *i* in *sample*
parameterized by the values of *mu* and *sigma* at index *i*.
}
\details{
Examples::

    sample = [[-2, -1, 0, 1, 2]]
    random_pdf_normal(sample=sample, mu=[0], sigma=[1]) =
        [[0.05399097, 0.24197073, 0.3989423, 0.24197073, 0.05399097]]

    random_pdf_normal(sample=sample*2, mu=[0,0], sigma=[1,2]) =
        [[0.05399097, 0.24197073, 0.3989423,  0.24197073, 0.05399097],
         [0.12098537, 0.17603266, 0.19947115, 0.17603266, 0.12098537]]


Defined in src/operator/random/pdf_op.cc:L300
}
