% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.reciprocal}
\alias{mx.symbol.reciprocal}
\title{reciprocal:Returns the reciprocal of the argument, element-wise.}
\usage{
mx.symbol.reciprocal(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Calculates 1/x.
}
\details{
Example::

reciprocal([-2, 1, 3, 1.6, 0.2]) = [-0.5, 1.0, 0.33333334, 0.625, 5.0]



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L638
}

