% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.reset_arrays}
\alias{mx.symbol.reset_arrays}
\title{reset_arrays:Set to zero multiple arrays}
\usage{
mx.symbol.reset_arrays(...)
}
\arguments{
\item{data}{NDArray-or-Symbol[]
Arrays}

\item{num.arrays}{int, required
number of input arrays.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Defined in src/operator/contrib/reset_arrays.cc:L36
}
