% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.sample_negative_binomial}
\alias{mx.symbol.sample_negative_binomial}
\title{sample_negative_binomial:Concurrent sampling from multiple
negative binomial distributions with parameters *k* (failure limit) and *p* (failure probability).}
\usage{
mx.symbol.sample_negative_binomial(...)
}
\arguments{
\item{k}{NDArray-or-Symbol
Limits of unsuccessful experiments.}

\item{shape}{Shape(tuple), optional, default=[]
Shape to be sampled from each random distribution.}

\item{dtype}{{'None', 'float16', 'float32', 'float64'},optional, default='None'
DType of the output in case this can't be inferred. Defaults to float32 if not defined (dtype=None).}

\item{p}{NDArray-or-Symbol
Failure probabilities in each experiment.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
The parameters of the distributions are provided as input arrays.
Let *[s]* be the shape of the input arrays, *n* be the dimension of *[s]*, *[t]*
be the shape specified as the parameter of the operator, and *m* be the dimension
of *[t]*. Then the output will be a *(n+m)*-dimensional array with shape *[s]x[t]*.
}
\details{
For any valid *n*-dimensional index *i* with respect to the input arrays, *output[i]*
will be an *m*-dimensional array that holds randomly drawn samples from the distribution
which is parameterized by the input values at index *i*. If the shape parameter of the
operator is not set, then one sample will be drawn per distribution and the output array
has the same shape as the input arrays.

Samples will always be returned as a floating point data type.

Examples::

k = [ 20, 49 ]
   p = [ 0.4 , 0.77 ]

// Draw a single sample for each distribution
   sample_negative_binomial(k, p) = [ 15.,  16.]

// Draw a vector containing two samples for each distribution
   sample_negative_binomial(k, p, shape=(2)) = [[ 15.,  50.],
                                                [ 16.,  12.]]

Defined in src/operator/random/multisample_op.cc:L289
}

