% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.slice_axis}
\alias{mx.symbol.slice_axis}
\title{slice_axis:Slices along a given axis.}
\usage{
mx.symbol.slice_axis(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
Source input}

\item{axis}{int, required
Axis along which to be sliced, supports negative indexes.}

\item{begin}{int, required
The beginning index along the axis to be sliced,  supports negative indexes.}

\item{end}{int or None, required
The ending index along the axis to be sliced,  supports negative indexes.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Returns an array slice along a given `axis` starting from the `begin` index
to the `end` index.
}
\details{
Examples::

x = [[  1.,   2.,   3.,   4.],
       [  5.,   6.,   7.,   8.],
       [  9.,  10.,  11.,  12.]]

slice_axis(x, axis=0, begin=1, end=3) = [[  5.,   6.,   7.,   8.],
                                           [  9.,  10.,  11.,  12.]]

slice_axis(x, axis=1, begin=0, end=2) = [[  1.,   2.],
                                           [  5.,   6.],
                                           [  9.,  10.]]

slice_axis(x, axis=1, begin=-3, end=-1) = [[  2.,   3.],
                                             [  6.,   7.],
                                             [ 10.,  11.]]

Defined in src/operator/tensor/matrix_op.cc:L501
}

