% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.smooth_l1}
\alias{mx.symbol.smooth_l1}
\title{smooth_l1:Calculate Smooth L1 Loss(lhs, scalar) by summing}
\usage{
mx.symbol.smooth_l1(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
source input}

\item{scalar}{float
scalar input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
.. math::
}
\details{
f(x) =
    \begin{cases}
    (\sigma x)^2/2,& \text{if }x < 1/\sigma^2\\
    |x|-0.5/\sigma^2,& \text{otherwise}
    \end{cases}

where :math:`x` is an element of the tensor *lhs* and :math:`\sigma` is the scalar.

Example::

smooth_l1([1, 2, 3, 4]) = [0.5, 1.5, 2.5, 3.5]
  smooth_l1([1, 2, 3, 4], scalar=1) = [0.5, 1.5, 2.5, 3.5]



Defined in src/operator/tensor/elemwise_binary_scalar_op_extended.cc:L104
}

