% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.softsign}
\alias{mx.symbol.softsign}
\title{softsign:Computes softsign of x element-wise.}
\usage{
mx.symbol.softsign(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
.. math::
   y = x / (1 + abs(x))
}
\details{
The storage type of ``softsign`` output is always dense



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L145
}

