% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.SpatialTransformer}
\alias{mx.symbol.SpatialTransformer}
\title{SpatialTransformer:Applies a spatial transformer to input feature map.}
\usage{
mx.symbol.SpatialTransformer(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
Input data to the SpatialTransformerOp.}

\item{loc}{NDArray-or-Symbol
localisation net, the output dim should be 6 when transform_type is affine. You shold initialize the weight and bias with identity tranform.}

\item{target.shape}{Shape(tuple), optional, default=[0,0]
output shape(h, w) of spatial transformer: (y, x)}

\item{transform.type}{{'affine'}, required
transformation type}

\item{sampler.type}{{'bilinear'}, required
sampling type}

\item{cudnn.off}{boolean or None, optional, default=None
whether to turn cudnn off}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
SpatialTransformer:Applies a spatial transformer to input feature map.
}

