% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.square}
\alias{mx.symbol.square}
\title{square:Returns element-wise squared value of the input.}
\usage{
mx.symbol.square(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
.. math::
   square(x) = x^2
}
\details{
Example::

square([2, 3, 4]) = [4, 9, 16]

The storage type of ``square`` output depends upon the input storage type:

- square(default) = default
   - square(row_sparse) = row_sparse
   - square(csr) = csr



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L815
}

