% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.stack}
\alias{mx.symbol.stack}
\title{stack:Join a sequence of arrays along a new axis.}
\usage{
mx.symbol.stack(...)
}
\arguments{
\item{data}{NDArray-or-Symbol[]
List of arrays to stack}

\item{axis}{int, optional, default='0'
The axis in the result array along which the input arrays are stacked.}

\item{num.args}{int, required
Number of inputs to be stacked.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
The axis parameter specifies the index of the new axis in the dimensions of the
result. For example, if axis=0 it will be the first dimension and if axis=-1 it
will be the last dimension.
}
\details{
Examples::

x = [1, 2]
  y = [3, 4]

stack(x, y) = [[1, 2],
                 [3, 4]]
  stack(x, y, axis=1) = [[1, 3],
                         [2, 4]]
}

