% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.tanh}
\alias{mx.symbol.tanh}
\title{tanh:Returns the hyperbolic tangent of the input array, computed element-wise.}
\usage{
mx.symbol.tanh(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input array.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
.. math::
   tanh(x) = sinh(x) / cosh(x)
}
\details{
The storage type of ``tanh`` output depends upon the input storage type:

- tanh(default) = default
   - tanh(row_sparse) = row_sparse
   - tanh(csr) = csr



Defined in src/operator/tensor/elemwise_unary_op_trig.cc:L234
}

