% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.transpose}
\alias{mx.symbol.transpose}
\title{transpose:Permutes the dimensions of an array.}
\usage{
mx.symbol.transpose(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
Source input}

\item{axes}{Shape(tuple), optional, default=[]
Target axis order. By default the axes will be inverted.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Examples::
}
\details{
x = [[ 1, 2],
       [ 3, 4]]

transpose(x) = [[ 1.,  3.],
                  [ 2.,  4.]]

x = [[[ 1.,  2.],
        [ 3.,  4.]],

[[ 5.,  6.],
        [ 7.,  8.]]]

transpose(x) = [[[ 1.,  5.],
                   [ 3.,  7.]],

[[ 2.,  6.],
                   [ 4.,  8.]]]

transpose(x, axes=(1,0,2)) = [[[ 1.,  2.],
                                 [ 5.,  6.]],

[[ 3.,  4.],
                                 [ 7.,  8.]]]

Defined in src/operator/tensor/matrix_op.cc:L312
}

