% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.zeros_like}
\alias{mx.symbol.zeros_like}
\title{zeros_like:Return an array of zeros with the same shape, type and storage type
as the input array.}
\usage{
mx.symbol.zeros_like(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
The input}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
The storage type of ``zeros_like`` output depends on the storage type of the input
}
\details{
- zeros_like(row_sparse) = row_sparse
- zeros_like(csr) = csr
- zeros_like(default) = default

Examples::

x = [[ 1.,  1.,  1.],
       [ 1.,  1.,  1.]]

zeros_like(x) = [[ 0.,  0.,  0.],
                   [ 0.,  0.,  0.]]
}

