% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mx.unserialize}
\alias{mx.unserialize}
\title{Unserialize MXNet model from Robject.}
\usage{
mx.unserialize(model)
}
\arguments{
\item{model}{The mxnet model loaded from RData files.}
}
\description{
Unserialize MXNet model from Robject.
}

