% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_url.R
\name{output_url}
\alias{output_url}
\title{Build URL for output file hosted on THREDDS fileServer}
\usage{
output_url(workflow_id, target, hostname = "localhost", port = 8000,
  https = FALSE)
}
\arguments{
\item{workflow_id}{ID of target workflow (numeric or character)}

\item{target}{Target path, relative to workflow directory (character)}

\item{hostname}{THREDDS server hostname (default = "localhost")}

\item{port}{THREDDS server port (default = 8000)}

\item{https}{Logical. If \code{TRUE}, use https, otherwise use http
(default = \code{FALSE}).}
}
\value{
THREDDS \code{http} \code{fileServer} URL (character)
}
\description{
Build URL for output file hosted on THREDDS fileServer
}
\author{
Alexey Shiklomanov
}
