% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_query.R
\name{param_query}
\alias{param_query}
\title{Run prepared query}
\usage{
param_query(con, query, params)
}
\arguments{
\item{con}{Database connection, as created by \link[RPostgres:dbConnect]{RPostgres::dbConnect}}

\item{query}{Query template (character, length 1)}

\item{params}{Query parameters (unnamed list)}
}
\value{
Query result, as \code{data.frame}
}
\description{
Run prepared query
}
\author{
Alexey Shiklomanov
}
