% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_query.R
\name{param_statement}
\alias{param_statement}
\title{Run prepared statement (e.g. \code{insert}, \code{delete}, \code{update})}
\usage{
param_statement(con, query, params)
}
\arguments{
\item{con}{Database connection, as created by \link[RPostgres:dbConnect]{RPostgres::dbConnect}}

\item{query}{Query template (character, length 1)}

\item{params}{Query parameters (unnamed list)}
}
\value{
Summary of operation (output of \link[DBI:dbBind]{DBI::dbBind})
}
\description{
Run prepared statement (e.g. \code{insert}, \code{delete}, \code{update})
}
\author{
Alexey Shiklomanov
}
