% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{PBMC Dataset}
\format{
A Seurat object
}
\source{
Example dataset for seuratAtlas tutorials
}
\usage{
pbmc
}
\description{
A Seurat object containing Peripheral Blood Mononuclear Cells (PBMC) data
for demonstration purposes.
}
\examples{
\dontrun{
data(pbmc)
pbmc
}
}
\keyword{datasets}
