% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_table}
\alias{pdf_table}
\title{Build a nice tibble for pdf}
\usage{
pdf_table(obj)
}
\arguments{
\item{obj}{is a tibble full of questions and responses, typically
from \code{diagramAPI::join_questions_responses}}
}
\description{
Builds a nicely formatted tibble of questions and responses
for a model (not scenario). Requires full set of simple_response
nodes.
}
