% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pecanapi_options}
\alias{pecanapi_options}
\title{Package-specific options}
\description{
To minimize the number of changes that have to happen for scripts
using \code{pecanapi} to be shared across machines, most
Docker/RabbitMQ/database-related configurations can be configured
via \code{options}. All options take the form \code{pecanapi.optionname},
where \code{optionname} is the specific option. Note that these values
are used only as default function arguments, and can be substituted
for any individual function call by passing the appropriate argument.
}
\details{
The following options (prefixed with \code{pecanapi.}) are used by
\code{pecanapi}. To see the default values, just call \code{options()} with
no arguments. These are sorted in order of decreasing likelihood of
needing to be set by the user (first options are most likely to be
changed across different systems):

-\code{pecanapi.user_id} -- The User ID to associate with all workflows
created by this package. This is the only option that \emph{must} be set
by the user -- it is set to \code{NULL} by default, which will cause
many of the functions in the \code{pecanapi} to fail.
\itemize{
\item \code{docker_hostname}, \code{docker_port} -- The hostname and port of the
Docker service. You can check that these values work by browsing to
\code{docker_hostname:docker_port} (by default, \code{localhost:8000}) in a
web browser.
\item \code{docker_rabbitmq_frontend} -- The "frontend rule" for RabbitMQ.
By default, this is \verb{/rabbitmq}, meaning that the RabbitMQ console
is accessible at \code{localhost:8000/rabbitmq} (adjusted for whatever
combination of \code{docker_hostname} and \code{docker_port} you are using).
}

\code{docker_https} -- (Logical) If \code{TRUE}, all URLs use \code{https} access.
By default, this is \code{FALSE}.
\itemize{
\item \code{db_hostname} -- The name of the PostgreSQL container service
inside the PEcAn stack. This is the same as its service name in
\code{docker-compose.yml}. This is the hostname used by the \code{executor}
service to access the database, and which is written into each
\code{pecan.xml} file.
\item \code{db_user}, \code{db_password}, \code{db_dbname}, \code{db_driver}, \code{db_write} --
These correspond to the \code{user}, \code{password}, \code{dbname}, \code{driver}, and
\code{write} tags in the \code{database/bety} part of the PEcAn XML.
\item \code{rabbitmq_user}, \code{rabbitmq_password} -- The RabbitMQ
authentication credentials. These are set in the
\code{docker-compose.yml} file, under the \code{rabbitmq} service.
\item \code{rabbitmq_service}, \code{rabbitmq_service_port}, \code{rabbitmq_vhost} --
The name, internal port, and \code{vhost} of the RabbitMQ service.
Unless you are making major changes to the guts of
\code{docker-compose.yml}, you shouldn't change these values (i.e. they
should be the same on most machines).
\item \code{workflow_hostname} -- The hostname passed to the \code{host} section
of the \code{pecan.xml}. By default, this is "docker".
\item \code{workflow_prefix} -- The location and directory prefix for
storing workflow outputs. By default, this is
\verb{/data/workflows/PEcAn_}. The workflow ID will be appended directly
to this value.
}
}
