% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Function to be executed when /ping API endpoint is called}
\usage{
ping()
}
\value{
Mapping containing response as "pong"
}
\description{
If successful connection to API server is established, this function will return the "pong" message
}
\author{
Tezan Sahu
}
