% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{process}
\alias{process}
\title{Complete sentiment analysis pipeline}
\usage{
process(page_name = NULL, project = NULL, language = NULL, api = NULL,
  format = "list", lexicon = NULL)
}
\arguments{
\item{page_name}{the title of the page you want to retrieve}

\item{project}{project you wish to query, if appropriate (e.g. "wikiquote")}

\item{language}{language code of the project you wish to query,
if appropriate (e.g. "de" for German)}

\item{api}{as an alternative to a \code{language} and \code{project} combination
(e.g. "wiki.mozilla.org/api.php" or "www.gutenberg.org/w/api.php")}

\item{format}{"json" when used in an endpoint or "list" (default) if using
this function in R}

\item{lexicon}{lexicon object provided by \link[tidytext:get_sentiments]{tidytext::get_sentiments}}
}
\description{
Outputs sentiment breakdown of (currently English) comments
either as a list for use in R or as a JSON string for use as an endpoint.
}
