% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_workflow.R
\name{reset_workflow_id_seq}
\alias{reset_workflow_id_seq}
\title{Reset workflow ID sequence to ID of last workflow in \code{workflows} table.}
\usage{
reset_workflow_id_seq(con)
}
\arguments{
\item{con}{Database connection, as created by \link[RPostgres:dbConnect]{RPostgres::dbConnect}}
}
\value{
Workflow ID of last workflow (64bit integer)
}
\description{
Reset workflow ID sequence to ID of last workflow in \code{workflows} table.
}
\author{
Alexey Shiklomanov
}
