% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnn.graph.R
\name{rnn.graph}
\alias{rnn.graph}
\title{Generate a RNN symbolic model - requires CUDA}
\usage{
rnn.graph(num_rnn_layer, input_size = NULL, num_embed = NULL, num_hidden,
  num_decode, dropout = 0, ignore_label = -1, bidirectional = F,
  loss_output = NULL, config, cell_type, masking = F,
  output_last_state = F, rnn.state = NULL, rnn.state.cell = NULL,
  prefix = "")
}
\arguments{
\item{num_rnn_layer}{int, number of stacked layers}

\item{input_size}{int, number of levels in the data - only used for embedding}

\item{num_embed}{int, default = NULL - no embedding. Dimension of the embedding vectors}

\item{num_hidden}{int, size of the state in each RNN layer}

\item{num_decode}{int, number of output variables in the decoding layer}

\item{dropout}{}

\item{config}{Either seq-to-one or one-to-one}

\item{cell_type}{Type of RNN cell: either gru or lstm}
}
\description{
Generate a RNN symbolic model - requires CUDA
}

