% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds_opendap.R
\name{run_dap}
\alias{run_dap}
\title{Create a THREDDS OpenDAP access URL to a specific model output file}
\usage{
run_dap(workflow_id, target, run_id = NULL, ...)
}
\arguments{
\item{workflow_id}{ID of target workflow (numeric or character)}

\item{target}{Target file path, relative to output directory of
specific run (as specified by \code{run_id})}

\item{run_id}{Run ID (numeric or character). If \code{NULL}, try to use
the run listed in the \code{runs.txt} file. If multiple runs are
available, throw a warning and use the first one.}

\item{...}{Additional arguments to \link{thredds_fs_url}}
}
\value{
OpenDAP URL to target file (character)
}
\description{
Create a THREDDS OpenDAP access URL to a specific model output file
}
\author{
Alexey Shiklomanov
}
