% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_url.R
\name{run_url}
\alias{run_url}
\title{Download a file from a specific}
\usage{
run_url(workflow_id, target, run_id = NULL, ...)
}
\arguments{
\item{workflow_id}{ID of target workflow (numeric or character)}

\item{target}{Target file path, relative to output directory of
specific run (as specified by \code{run_id})}

\item{run_id}{Run ID (numeric or character). If \code{NULL}, try to use
the run listed in the \code{runs.txt} file. If multiple runs are
available, throw a warning and use the first one.}

\item{...}{Additional arguments to \link{output_url}}
}
\value{
HTTP fileServer URL of a particular run output file (character)
}
\description{
Download a file from a specific
}
\author{
Alexey Shiklomanov
}
