% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_models}
\alias{search_models}
\alias{search_sites}
\alias{search_pfts}
\title{Search for sites or models}
\usage{
search_models(
  con,
  name = "",
  revision = "",
  modeltype = "",
  auto_pct = TRUE,
  ignore.case = TRUE
)

search_sites(con, sitename = "", auto_pct = TRUE, ignore.case = TRUE)

search_pfts(
  con,
  name = "",
  definition = "",
  modeltype = "",
  auto_pct = TRUE,
  ignore.case = TRUE
)
}
\arguments{
\item{con}{Database connection, as created by \link[RPostgres:dbConnect]{RPostgres::dbConnect}}

\item{name}{Model/PFT (depending on function) name search string (character)}

\item{revision}{Model version search string (character)}

\item{modeltype}{Model type search string (character)}

\item{auto_pct}{Logical. If \code{TRUE} (default), automatically
surround search strings in \verb{\%}. If this is \code{FALSE}, you should
explicitly specify \code{"\%\%"} for one or both other arguments.}

\item{ignore.case}{Logical. If \code{TRUE} (default) use
case-insensitive search (SQL \code{ILIKE} operator); otherwise, use
case-sensitive search (SQL \code{LIKE} operator).}

\item{sitename}{Model name search string (character)}

\item{definition}{PFT definition search string (character)}
}
\value{
Bety \code{models} table (\code{data.frame}) subset to matching model
name or version
}
\description{
Search for sites or models
}
\examples{
\dontrun{
search_models(con, "SIPNET")

# Partial match
search_models(con, "ED")
search_models(con, modeltype = "ED")
search_sites(con, "UMBS")
search_pfts(con, "early", modeltype = "ED")

# Case sensitivity
search_models(con, "ed")
search_models(con, "ed", ignore.case = FALSE)

# Starts with UMBS
search_sites(con, "UMBS\%", auto_pct = FALSE)

# SQL wildcards can still be used inside search strings.
search_pfts(con, "early\%hardwood")
}
}
\author{
Alexey Shiklomanov
}
