% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentalk.R
\docType{package}
\name{sentimentalk-package}
\alias{sentimentalk-package}
\alias{sentimentalk}
\title{sentimentalk: Sentiment Analysis of Talk Pages as Service}
\description{
\href{https://plumber.trestletech.com/}{plumber}-powered API for
\href{http://tidytextmining.com/sentiment.html}{tidy sentiment analysis} of
\href{https://www.mediawiki.org/wiki/Help:Talk_pages}{MediaWiki talk pages} on
\href{https://www.mediawiki.org/wiki/MediaWiki}{MediaWiki}-powered websites in
R.
}
\references{
\itemize{
\item \emph{\href{http://tidytextmining.com/}{Text Mining with R: A Tidy Approach}}
by Julia Silge and David Robinson
\item Mohammad, S. M., & Yang, T. W. (2011, June). Tracking sentiment in mail:
how genders differ on emotional axes. In Proceedings of the 2nd workshop
on computational approaches to subjectivity and sentiment analysis
(pp. 70-79). Association for Computational Linguistics.
}
}
