% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_workflow.R
\name{set_workflow_id_seq}
\alias{set_workflow_id_seq}
\title{Set workflow ID sequence to a specific value}
\usage{
set_workflow_id_seq(con, value)
}
\arguments{
\item{con}{Database connection, as created by \link[RPostgres:dbConnect]{RPostgres::dbConnect}}

\item{value}{Workflow ID to which to reset workflow ID sequence (base64 integer)}
}
\value{
Input \code{value}
}
\description{
Set workflow ID sequence to a specific value
}
\author{
Alexey Shiklomanov
}
