% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_workflow.R
\name{submit_workflow}
\alias{submit_workflow}
\title{Post complete settings list as RabbitMQ message}
\usage{
submit_workflow(settings, rabbitmq_hostname = "localhost",
  rabbitmq_frontend = "/rabbitmq", rabbitmq_port = 8000,
  rabbitmq_user = "guest", rabbitmq_password = "guest",
  rabbitmq_prefix = NULL, httr_auth = TRUE, https = FALSE)
}
\arguments{
\item{settings}{PEcAn settings list object}

\item{rabbitmq_hostname}{RabbitMQ server hostname (character.
Default = \code{"localhost"})}

\item{rabbitmq_port}{RabbitMQ server port (character or numeric.
Default = 8000)}

\item{rabbitmq_user}{RabbitMQ user name (character. Default = \code{"guest"})}

\item{rabbitmq_password}{RabbitMQ password (character. Default = \code{"guest"})}

\item{rabbitmq_prefix}{Complete RabbitMQ API prefix. If \code{NULL}
(default), this is constructed from the other arguments. If this
argument is not \code{NULL}, it overrides all other arguments except
\code{httr_auth} and \code{settings}.}

\item{httr_auth}{Whether or not to use \link[httr:authenticate]{httr::authenticate} to
generate CURL authentication (default = \code{TRUE}). If \code{FALSE}, you
must pass the authentication as part of the RabbitMQ hostname or prefix.}

\item{https}{Whether or not to use \code{https}. If \code{FALSE} (default),
use \code{http}.}
}
\value{
Curl \code{POST} output, parsed by \link[httr:content]{httr::content}
}
\description{
Post complete settings list as RabbitMQ message
}
\author{
Alexey Shiklomanov
}
