% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds_opendap.R
\name{thredds_dap_url}
\alias{thredds_dap_url}
\title{Create a THREDDS OpenDAP access URL to a PEcAn file}
\usage{
thredds_dap_url(
  target,
  hostname = getOption("pecanapi.docker_hostname"),
  port = getOption("pecanapi.docker_port"),
  https = getOption("pecanapi.docker_https")
)
}
\arguments{
\item{target}{Full path to target (character) relative to THREDDS
root. For outputs, this should start with "outputs/", and for
dbfiles, "dbfiles/".}

\item{hostname}{THREDDS server hostname (default = "localhost")}

\item{port}{THREDDS server port (default = 8000)}

\item{https}{Logical. If \code{TRUE}, use https, otherwise use http
(default = \code{FALSE}).}
}
\value{
OpenDAP URL to target file (character)
}
\description{
Create a THREDDS OpenDAP access URL to a PEcAn file
}
\author{
Alexey Shiklomanov
}
