% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds_fileserver.R
\name{thredds_fs_url}
\alias{thredds_fs_url}
\title{Create a THREDDS fileServer URL prefix}
\usage{
thredds_fs_url(hostname = "localhost", port = 8000, https = FALSE)
}
\arguments{
\item{hostname}{THREDDS server hostname (default = "localhost")}

\item{port}{THREDDS server port (default = 8000)}

\item{https}{Logical. If \code{TRUE}, use https, otherwise use http
(default = \code{FALSE}).}
}
\value{
THREDDS fileServer URL prefix (character)
}
\description{
Create a THREDDS fileServer URL prefix
}
\author{
Alexey Shiklomanov
}
