% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_numeric.R
\name{to_numeric_generator}
\alias{to_numeric_generator}
\title{To numeric generator}
\usage{
to_numeric_generator(class_name = NULL)
}
\arguments{
\item{class_name}{class attribute to give a single response
node that will be passed to the generic to_numeric}
}
\value{
a function that takes a single response node, attaches
the class_name as a class and calls to_numeric
}
\description{
Utility function for creating to_numeric functions
for each node in a response.
}
