% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{unpack_json}
\alias{unpack_json}
\title{Unpack json}
\usage{
unpack_json(res)
}
\arguments{
\item{res}{a list (of lists)}
}
\value{
a list (of lists) where unnamed lists have become vectors.
}
\description{
Convert unnamed lists in a list of lists to a vector.
This ensures that JSON arrays are converted to vectors rather than
lists. Operates recursively across the whole list of lists.
}
