% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch_progress.R
\name{watch_workflow}
\alias{watch_workflow}
\title{Continuously monitor the progress of a workflow until it completes.
To exit early, send an interrupt signal (Control-C).}
\usage{
watch_workflow(workflow_id, nlines = 10, sleep = 3, ...)
}
\arguments{
\item{workflow_id}{Workflow ID (character or numeric)}

\item{nlines}{Print the last N lines (numeric, default = 10)}

\item{sleep}{Number of seconds to sleep between status updates
(numeric, default = 3)}

\item{...}{Additional arguments to \link{output_url}}
}
\value{
If successful
}
\description{
Continuously monitor the progress of a workflow until it completes.
To exit early, send an interrupt signal (Control-C).
}
\author{
Alexey Shiklomanov
}
