% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch_progress.R
\name{workflow_output}
\alias{workflow_output}
\title{Retrieve the current output of a workflow given its ID}
\usage{
workflow_output(workflow_id, ...)
}
\arguments{
\item{workflow_id}{Workflow ID (character or numeric)}

\item{...}{Additional arguments to \link{output_url}}
}
\value{
Workflow output, as character vector with one item per line
in \code{workflow.Rout} (as returned by \code{readLines}).
}
\description{
Retrieve the current output of a workflow given its ID
}
\author{
Alexey Shiklomanov
}
