% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Perform a sentiment analysis of a parsed, tidy talk page}
\usage{
analyze(tidy_page, lexicon)
}
\arguments{
\item{tidy_page}{the output from \link{parse_discussion}}

\item{lexicon}{lexicon object provided by \link[tidytext:get_sentiments]{tidytext::get_sentiments}}
}
\description{
Uses an emotion lexicon to output a breakdown of sentiments
expressed.
}
