% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_workflow.R
\name{get_next_workflow_id}
\alias{get_next_workflow_id}
\title{Get current workflow ID and update internal workflow ID PostgreSQL
sequence}
\usage{
get_next_workflow_id(con)
}
\arguments{
\item{con}{Database connection, as created by \link[RPostgres:dbConnect]{RPostgres::dbConnect}}
}
\value{
Workflow ID, as numeric/base64 integer
}
\description{
The \code{workflows} table has an internal
\href{https://www.postgresql.org/docs/9.6/sql-createsequence.html}{sequence}
that keeps track of and automatically updates the workflow ID
(that's why inserting into the table without explicitly setting a
workflow ID is a safe and robust operation). This function is a
wrapper around the
\href{https://www.postgresql.org/docs/9.6/functions-sequence.html}{\code{nextval} function},
which retrieves the current value of the sequence \emph{and} augments
the sequence by 1.
}
\author{
Alexey Shiklomanov
}
