% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_workflow.R
\name{insert_new_workflow}
\alias{insert_new_workflow}
\title{Insert a new workflow into PEcAn database, returning the workflow
as a \code{data.frame}}
\usage{
insert_new_workflow(
  con,
  site_id,
  model_id,
  start_date,
  end_date,
  user_id = getOption("pecanapi.user_id"),
  hostname = getOption("pecanapi.workflow_hostname"),
  folder_prefix = getOption("pecanapi.workflow_prefix"),
  params = NULL,
  notes = NULL
)
}
\arguments{
\item{con}{Database connection, as created by \link[RPostgres:dbConnect]{RPostgres::dbConnect}}

\item{site_id}{Site ID from \code{sites} table (numeric)}

\item{model_id}{Model ID from \code{models} table (numeric)}

\item{start_date}{Model run start date (character or POSIX)}

\item{end_date}{Model run end date (character or POSIX)}

\item{user_id}{User ID from \code{users} table (default = option
\code{pecanapi.user_id}). Note that this option is \emph{not set by
default}, and this function will not run without a set \code{user_id}.}

\item{hostname}{Workflow server hostname (character; default =
option \code{pecanapi.workflow_hostname})}

\item{folder_prefix}{Output directory prefix (character; default =
option \code{pecanapi.workflow_prefix}). Workflow ID will be appended
to the end with \code{paste0}}

\item{params}{Additional workflow parameters, stored in
\code{workflows.params} (character or NULL (default))}

\item{notes}{Additional workflow notes, stored in \code{workflows.notes}
(character or NULL (default))}
}
\value{
\code{data.frame} containing new workflow(s), including all
columns from \code{workflows} table.
}
\description{
Insert a new workflow into PEcAn database, returning the workflow
as a \code{data.frame}
}
\author{
Alexey Shiklomanov
}
