% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.nd.IdentityAttachKLSparseReg}
\alias{mx.nd.IdentityAttachKLSparseReg}
\title{Apply a sparse regularization to the output a sigmoid activation function.}
\arguments{
\item{data}{NDArray-or-Symbol
Input data.}

\item{sparseness.target}{float, optional, default=0.1
The sparseness target}

\item{penalty}{float, optional, default=0.001
The tradeoff parameter for the sparseness penalty}

\item{momentum}{float, optional, default=0.9
The momentum for running average}
}
\value{
out The result mx.ndarray
}
\description{
Apply a sparse regularization to the output a sigmoid activation function.
}

