% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxnet_generated.R
\name{mx.symbol.shape_array}
\alias{mx.symbol.shape_array}
\title{shape_array:Returns a 1D int64 array containing the shape of data.}
\usage{
mx.symbol.shape_array(...)
}
\arguments{
\item{data}{NDArray-or-Symbol
Input Array.}

\item{lhs.begin}{int or None, optional, default='None'
Defaults to 0. The beginning index along which the lhs dimensions are to be reshaped. Supports negative indices.}

\item{lhs.end}{int or None, optional, default='None'
Defaults to None. The ending index along which the lhs dimensions are to be used for reshaping. Supports negative indices.}

\item{rhs.begin}{int or None, optional, default='None'
Defaults to 0. The beginning index along which the rhs dimensions are to be used for reshaping. Supports negative indices.}

\item{rhs.end}{int or None, optional, default='None'
Defaults to None. The ending index along which the rhs dimensions are to be used for reshaping. Supports negative indices.}

\item{name}{string, optional
Name of the resulting symbol.}
}
\value{
out The result mx.symbol
}
\description{
Example::
}
\details{
shape_array([[1,2,3,4], [5,6,7,8]]) = [2,4]



Defined in src/operator/tensor/elemwise_unary_op_basic.cc:L505
}

