% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds_fileserver.R
\name{output_url}
\alias{output_url}
\title{Build URL for output file hosted on THREDDS fileServer}
\usage{
output_url(workflow_id, target, ...)
}
\arguments{
\item{workflow_id}{ID of target workflow (numeric or character)}

\item{target}{Target path, relative to workflow directory (character)}

\item{...}{Additional arguments to \link{thredds_fs_url}}
}
\value{
THREDDS http fileServer URL (character)
}
\description{
Build URL for output file hosted on THREDDS fileServer
}
\author{
Alexey Shiklomanov
}
