% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_discussion}
\alias{parse_discussion}
\title{Parse a downloaded talk page}
\usage{
parse_discussion(talk_page)
}
\arguments{
\item{talk_page}{the talk page HTML code downloaded using
\link[WikipediR:page_content]{WikipediR::page_content}}
}
\description{
Parses a downloaded talk page into a tidy format with
identifiers for topic and authors.
}
